### 2022-03-21 #4259 @danielskatz
👋 @ebezzam - this software looks really interesting, but I'm not sure if it fits the definition of [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS. This does not mean that it is not software that is useful in research, but just that JOSS might not consider it in scope for review as research software.

The editors will now discuss if it meets this criterion for review by JOSS.  You should hear back in a week or two.
