### 2022-03-21 #4259 @xuanxu
The [Crossref API](https://api.crossref.org/) is currently down and returning timeout errors.
The `check references` command should work once they solve the problem.
