### 2022-03-21 #4259 @danielskatz
Is this something we can catch and report to the command issuer?
