### 2022-03-21 #4259 @xuanxu
> Is this something we can catch and report to the command issuer?
Probably, at least a general message pointing to a problem with the Crossref calls should be feasible
