+++ paper/paper.bib
@@ -1,93 +1,125 @@
 @article{boominathan2022recent,
-  title={Recent advances in lensless imaging},
-  author={Boominathan, Vivek and Robinson, Jacob T and Waller, Laura and Veeraraghavan, Ashok},
-  journal={Optica},
-  volume={9},
-  number={1},
-  pages={1--16},
-  year={2022},
-  publisher={Optical Society of America}
+    author = {Vivek Boominathan and Jacob T. Robinson and Laura Waller and Ashok Veeraraghavan},
+    journal = {Optica},
+    keywords = {Charge coupled devices; Image processing; Imaging systems; Three dimensional imaging; Vertical cavity surface emitting lasers; X ray imaging},
+    number = {1},
+    pages = {1--16},
+    publisher = {OSA},
+    title = {Recent advances in lensless imaging},
+    volume = {9},
+    month = {Jan},
+    year = {2022},
+    url = {http://opg.optica.org/optica/abstract.cfm?URI=optica-9-1-1},
+    doi = {10.1364/OPTICA.431361},
+    abstract = {Lensless imaging provides opportunities to design imaging systems free from the constraints imposed by traditional camera architectures. Due to advances in imaging hardware, fabrication techniques, and new algorithms, researchers have recently developed lensless imaging systems that are extremely compact and lightweight or able to image higher-dimensional quantities. Here we review these recent advances and describe the design principles and their effects that one should consider when developing and using lensless imaging systems.},
 }
 
-@inproceedings{bezzam2017hardware,
-  title={Hardware and software for reproducible research in audio array signal processing},
-  author={Bezzam, Eric and Scheibler, Robin and Azcarreta, Juan and Pan, Hanjie and Simeoni, Matthieu and Beuchat, Ren{\'e} and Hurley, Paul and Bruneau, Basile and Ferry, Corentin and Kashani, Sepand},
-  booktitle={2017 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
-  pages={6591--6592},
-  year={2017},
-  organization={Ieee}
+
+
+@INPROCEEDINGS{bezzam2017hardware,
+    author={Bezzam, Eric and Scheibler, Robin and Azcarreta, Juan and Pan, Hanjie and Simeoni, Matthieu and Beuchat, Rene and Hurley, Paul and Bruneau, Basile and Ferry, Corentin and Kashani, Sepand},
+    booktitle={2017 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
+    title={Hardware and software for reproducible research in audio array signal processing},
+    year={2017},
+    volume={},
+    number={},
+    pages={6591-6592},
+    doi={10.1109/ICASSP.2017.8005297}
 }
 
 @misc{diffusercam,
-  author = {C. Biscarrat and S. Parthasarathy and G. Kuo and N. Antipa},
-  title = {Build your own DiffuserCam: {T}utorial},
-  year = {2018},
-  url = {https://waller-lab.github.io/DiffuserCam/tutorial.html}
+    author = {C. Biscarrat and S. Parthasarathy and G. Kuo and N. Antipa},
+    title = {Build your own DiffuserCam: {T}utorial},
+    year = {2018},
+    url = {https://waller-lab.github.io/DiffuserCam/tutorial.html}
 }
 
-@inproceedings{bezzam2019teaching,
-  title={Teaching Practical DSP with Off-the-shelf Hardware and Free Software},
-  author={Bezzam, Eric and Hoffet, Adrien and Prandoni, Paolo},
-  booktitle={ICASSP 2019-2019 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
-  pages={7660--7664},
-  year={2019},
-  organization={IEEE}
[…]
