### 2022-04-01 #4259 @ebezzam
> but I'm not sure if it fits the definition of [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS.

@danielskatz thank you for raising this concern. To address this point, as the title may not reflect that this is actually a software package. Here's how we believe `LenslessPiCam` meets the scope of research software as defined by JOSS:

- “Solves complex modeling problems in in a scientific context”: the proposed package includes an algorithmic component with multiple techniques for solving problems of the form Eq 1 in the paper, which are typical in lensless imaging.
- “Supports the functioning of research instruments or the execution of research experiments”: there is very little software available online to conduct lensless imaging experiments (only one we are aware of is [DiffuserCam-Tutorial](https://github.com/Waller-Lab/DiffuserCam-Tutorial) which only provides a script for capture and with a poor quality sensor). We provide the designs to build a lensless camera and software to collect a dataset with it.

Moreover, the package does not require building the lensless camera in order to find use. The modular algorithmic components and functionality to compute performance metrics can be used with datasets collected by other cameras (as we've shown with [this popular dataset](https://waller-lab.github.io/LenslessLearning/dataset.html) in the lensless imaging community) in order to explore new reconstruction approaches.

Hope that helps clarify.
