### 2022-04-06 #4259 @Kevin-Mattheus-Moerman
@danasolav this may be a bit off topic but do you think with your experience with Raspberry Pi cameras you can help edit this work? I can help too if needed.
