### 2022-04-06 #4259 @Kevin-Mattheus-Moerman
@ebezzam the editorial board has determined this may be in scope for JOSS so I've removed the query-scope flag. We are now proceeding to find a handling editor.
