### 2022-04-11 a66f7d8
Add MNIST original data.
