### 2022-04-12 81005ee
Add grayscale and downsample for remote capture, reduce dependencies for RPi.
