### 2022-04-12 aca5208
Add grayscale and downsample for remote capture, reduce dependencies for RPi.
