### 2022-04-13 2cdb276
Add script for measuring MNIST dataset.
