+++ scripts/collect_mnist.py
@@ -1,0 +1,196 @@
+import torchvision.datasets as dset
+from torchvision import transforms
+import torch
+import numpy as np
+import time
+import os
+import subprocess
+import pathlib as plib
+import click
+
+
+@click.command()
+@click.option(
+    "--hostname",
+    type=str,
+    help="Hostname or IP address for display device.",
+)
+@click.option(
+    "--camera_hostname",
+    type=str,
+    help="Hostname or IP address for capture device. If not provided, same as hostname",
+)
+@click.option(
+    "--output_dir",
+    type=str,
+    help="Output directory for measured images.",
+)
+@click.option(
+    "--n_files",
+    type=int,
+    help="Number of files to collect. Default is all files.",
+)
+@click.option(
+    "--test",
+    is_flag=True,
+    help="Measure test set, otherwise do train.",
+)
+@click.option("-v", "--verbose", is_flag=True)
+def collect_mnist(hostname, camera_hostname, output_dir, n_files, verbose, test):
+
+    assert hostname is not None
+    assert output_dir is not None
+
+    # TODO : save metadata in JSON
+    progress = 10
+    downsample = 128
+
+    # display param
+    display_python = "~/LenslessPiCam/lensless_env/bin/python"
+    display_tmp_file = "~/tmp_display.png"
+    display_image_prep_script = "~/LenslessPiCam/scripts/prep_display_image.py"
+    display_image_path = "~/LenslessPiCam_display/test.png"
+    display_res = np.array((1920, 1200))
+    hshift = 0
+    vshift = 0
+    pad = 50
+    brightness = 100
+    local_display_file = "display_tmp.png"
+
+    # camera param
+    camera_python = "~/LenslessPiCam/lensless_env/bin/python"
+    if camera_hostname is None:
+        camera_hostname = hostname
+    sensor_mode = "0"
+    exp = 0.2
+    config_pause = 0
+    capture_script = "~/LenslessPiCam/scripts/on_device_capture.py"
+
+    # load datasets
+    root = "./data"
+    tr = None
+    # tr = transforms.ToTensor()
+    dataset = dset.MNIST(root=root, train=not test, download=True, transform=tr)
+
+    print("\nNumber of files :", len(dataset))
+    if n_files:
+        print(f"TEST : collecting first {n_files} files!")
+    else:
+        n_files = len(dataset)
+
+    # make output directory
+    output_dir = plib.Path(output_dir)
+    output_dir.mkdir(exist_ok=True)
+
+    # loop over files
+    if test:
+        subdir = output_dir / "test"
+    else:
+        subdir = output_dir / "train"
+    subdir.mkdir(exist_ok=True)
+    labels = []
+    start_time = time.time()
+    for i in range(n_files):
+
+        if verbose:
+            print(f"\nFILE : {i+1} / {n_files}")
+        img, label = dataset[i]
+        labels.append(label)
+
+        # TODO check if measurement already exists
+        output_fp = subdir / f"img{i}.png"
+        if not os.path.isfile(output_fp):
+
+            # load and save as PNG
+            img.save(local_display_file)
+
+            # send to display
+            if verbose:
+                print("-- Displaying picture...")
+            # -- copy to RPi
+            os.system('scp %s "pi@%s:%s" ' % (local_display_file, hostname, display_tmp_file))
+
+            # -- prep image on Pi
+            prep_command = f"{display_python} {display_image_prep_script} --fp {display_tmp_file} \
+                --pad {pad} --vshift {vshift} --hshift {hshift} --screen_res {display_res[0]} {display_res[1]} \
+                --brightness {brightness} --output_path {display_image_path} "
+            if verbose:
+                print(f"Command : {prep_command}")
+            subprocess.Popen(
[…]
