### 2022-04-13 eef79e7
Add script for measuring MNIST dataset.
