### 2022-04-21 14db41b
Add script to run MNIST collection on RPi.
