+++ scripts/collect_mnist_on_device.py
@@ -1,0 +1,181 @@
+import numpy as np
+import time
+import os
+import pathlib as plib
+import click
+import cv2
+from mlxtend.data import loadlocal_mnist
+from PIL import Image
+from picamerax import PiCamera
+
+
+@click.command()
+@click.option(
+    "--input_dir",
+    type=str,
+    default="data/MNIST",
+    help="Where raw MNIST data is stored.",
+)
+@click.option(
+    "--output_dir",
+    type=str,
+    help="Output directory for measured images.",
+)
+@click.option(
+    "--n_files",
+    type=int,
+    help="Number of files to collect. Default is all files.",
+)
+@click.option(
+    "--test",
+    is_flag=True,
+    help="Measure test set, otherwise do train.",
+)
+@click.option(
+    "--bayer",
+    is_flag=True,
+    help="Whether to save Bayer data (big!) or de-mosaiced and downsampled.",
+)
+@click.option(
+    "--downsample",
+    type=float,
+    default=128,
+    help="Amount to downsample.",
+)
+@click.option("-v", "--verbose", is_flag=True)
+def collect_mnist(input_dir, output_dir, n_files, verbose, test, bayer, downsample):
+
+    assert output_dir is not None
+
+    img_og_dim = (28, 28)
+    output_dim = (np.array([4056, 3040]) / downsample).astype(int)
+    interpolation = cv2.INTER_NEAREST
+
+    # TODO : save metadata in JSON
+    progress = 100
+
+    # display param
+    screen_res = np.array((1920, 1200))
+    hshift = 0
+    vshift = 0
+    pad = 50
+    brightness = 100
+    display_image_path = "/home/pi/LenslessPiCam_display/test.png"
+
+    # load data
+    if test:
+        X, y = loadlocal_mnist(
+            images_path=os.path.join(input_dir, "t10k-images-idx3-ubyte"),
+            labels_path=os.path.join(input_dir, "t10k-labels-idx1-ubyte"),
+        )
+    else:
+        X, y = loadlocal_mnist(
+            images_path=os.path.join(input_dir, "train-images-idx3-ubyte"),
+            labels_path=os.path.join(input_dir, "train-labels-idx1-ubyte"),
+        )
+
+    print("\nNumber of files :", len(y))
+    if n_files:
+        print(f"TEST : collecting first {n_files} files!")
+    else:
+        n_files = len(y)
+
+    # set up camera for consistent photos
+    # https://picamera.readthedocs.io/en/release-1.13/recipes1.html#capturing-consistent-images
+    # https://picamerax.readthedocs.io/en/latest/fov.html?highlight=camera%20resolution#sensor-modes
+    camera = PiCamera(resolution=(640, 480), framerate=30)
+    # camera = PiCamera(resolution=output_dim, framerate=30)
+    # Set ISO to the desired value
+    camera.iso = 100
+    # Wait for the automatic gain control to settle
+    time.sleep(2)
+    # Now fix the values
+    camera.shutter_speed = camera.exposure_speed
+    camera.exposure_mode = "off"
+    g = camera.awb_gains
+    camera.awb_mode = "off"
+    camera.awb_gains = g
+
+    # make output directory
+    output_dir = plib.Path(output_dir)
+    output_dir.mkdir(exist_ok=True)
+
+    # loop over files
+    if test:
+        subdir = output_dir / "test"
+    else:
+        subdir = output_dir / "train"
+    subdir.mkdir(exist_ok=True)
+    labels = []
+    start_time = time.time()
+    for i in range(n_files):
+
+        if verbose:
+            print(f"\nFILE : {i+1} / {n_files}")
+
+        labels.append(y[i])
+
+        # TODO check if measurement already exists
+        output_fp = subdir / f"img{i}.png"
+        if not os.path.isfile(output_fp):
+
[…]
