### 2022-04-21 de282d6
Save capture settings.
