### 2022-04-22 ad934d0
Download MNIST is not available.
