+++ scripts/collect_mnist_on_device.py
@@ -98,12 +98,12 @@ def collect_mnist(
         zipped_file = images_path + ".gz"
         response = requests.get(f"http://yann.lecun.com/exdb/mnist/{zipped_file}")
         open(zipped_file, "wb").write(response.content)
-        os.system("gunzip " + zipped_file + ".gz")
+        os.system("gunzip " + zipped_file)
     if not os.path.exists(labels_path):
         print("Downloading labels...")
         zipped_file = labels_path + ".gz"
         response = requests.get(f"http://yann.lecun.com/exdb/mnist/{zipped_file}")
-        os.system("gunzip " + zipped_file + ".gz")
+        os.system("gunzip " + zipped_file)
 
     X, y = loadlocal_mnist(
         images_path=images_path,
