+++ README.md
@@ -1,6 +1,12 @@
-# LenslessPiCam
+# LenslessPiCam : Lensless Imaging with a Raspberry Pi
 
-### _Lensless imaging with a Raspberry Pi and a piece of tape!_
+- [Setup](#setup).
+- [Data for examples](#data).
+- [Reconstruction](#recon).
+- [Evaluating on a dataset](#eval).
+- [Remote capture](#capture).
+- [Remote display](#display).
+- [Collecting MNIST](#mnist).
 
 ![Example PSF, raw data, and reconstruction.](scripts/example_reconstruction.png)
 
@@ -32,7 +38,7 @@ course at EPFL, and therefore includes some exercises / code to complete:
 If you are an instructor, you can request access to the solutions [here](https://drive.google.com/drive/folders/1Y1scM8wVfjVAo5-8Nr2VfE4b6VHeDSia?usp=sharing) 
 or send an email to `eric[dot]bezzam[at]epfl[dot]ch`.
 
-## Setup
+## Setup <a name="setup"></a>
 The expected workflow is to have a local computer which interfaces remotely
 with a Raspberry Pi equipped with the HQ camera sensor (or V2 sensor as in the 
 original tutorial).
@@ -89,7 +95,7 @@ and the Raspberry Pi. There are two possible solutions to this, as proposed in
 2. Comment `AcceptEnv LANG LC_*` in `/etc/ssh/sshd_config` on the Raspberry
 Pi.
 
-## Data for examples
+## Data for examples <a name="data"></a>
 
 You can download example PSFs and raw data that we've measured [here](https://drive.switch.ch/index.php/s/NdgHlcDeHVDH5ww).
 We recommend placing this content in the `data` folder.
@@ -98,7 +104,7 @@ You can download a subset for the [DiffuserCam Lensless Mirflickr Dataset](https
 that we've prepared [here](https://drive.switch.ch/index.php/s/vmAZzryGI8U8rcE)
 with `scripts/prepare_mirflickr_subset.py`.
 
-## Reconstruction
+## Reconstruction <a name="recon"></a>
 
 The core algorithmic component of `LenslessPiCam` is the abstract class 
 `lensless.ReconstructionAlgorithm`. The three reconstruction strategies 
@@ -140,7 +146,7 @@ For example, for ADMM (full example in `scripts/admm.py`):
 A template for applying a reconstruction algorithm (including loading the data)
 can be found in `scripts/reconstruction_template.py`.
 
-## Evaluating on a dataset
+## Evaluating on a dataset <a name="eval"></a>
 
 You can run ADMM on the [DiffuserCam Lensless Mirflickr Dataset](https://waller-lab.github.io/LenslessLearning/dataset.html)
 with the following script.
@@ -168,7 +174,7 @@ python scripts/apply_admm_single_mirflickr.py \
 --fid 172
 ```
 
-## Remote capture
+## Remote capture <a name="capture"></a>
 
 You can remotely capture raw Bayer data with the following script.
 ```bash
@@ -178,7 +184,7 @@ where `<HOSTNAME>` is the hostname or IP address of your Raspberry Pi, `<FN>` is
 the name of the file to save the Bayer data, and the other arguments can be used
 to adjust camera settings.
 
-## Remote display
+## Remote display <a name="display"></a>
 
 For collecting images displayed on a screen, we have prepared some software to
 remotely display images on a Raspberry Pi installed with this software and
@@ -211,16 +217,26 @@ the path on your local computer of the image you would like to display, and the
 other arguments can be used to adjust the positioning of the image and its
 brightness.
 
-## Formatting
-You can use `black` to format your code.
+## Collecting MNIST <a name="mnist"></a>
 
-First install the library.
+We provide a couple scripts to collect MNIST with the proposed camera.
+
[…]
