+++ scripts/recon/apgd_pycsou.py
@@ -1,22 +1,34 @@
 """
-Apply ADMM reconstruction.
+Apply Accelerated Proximal Gradient Descent (APDG) with a non-negativity prior
+for grayscale reconstruction. Example using Pycsou:
+https://matthieumeo.github.io/pycsou/html/api/algorithms/pycsou.opt.proxalgs.html?highlight=apgd#pycsou.opt.proxalgs.AcceleratedProximalGradientDescent
 
+Example
 ```
-python scripts/admm.py --psf_fp data/psf/tape_rgb.png \
---data_fp data/raw_data/thumbs_up_rgb.png --n_iter 5
+python scripts/apgd_pycsou.py --psf_fp data/psf/tape_rgb.png \
+--data_fp data/raw_data/thumbs_up_rgb.png --real_conv
 ```
+Note that the `RealFFTConvolve2D` has to be implemented in `lensless/realfftconv.py`.
+
+Otherwise, grayscale reconstruction with the non-optimized FFT convolution can
+be readily used (RGB is not supported):
+```
+python scripts/apgd_pycsou.py --psf_fp data/psf/tape_rgb.png --data_fp \
+data/raw_data/thumbs_up_rgb.png --gray
+```
+
 
 """
 
-import os
+import numpy as np
 import time
-import pathlib as plib
+from datetime import datetime
 import click
 import matplotlib.pyplot as plt
-from datetime import datetime
-import numpy as np
 from lensless.io import load_data
-from lensless import ADMM
+from lensless import APGD, APGDPriors
+import os
+import pathlib as plib
 
 
 @click.command()
@@ -31,10 +43,16 @@ from lensless import ADMM
     help="File name for raw measurement data.",
 )
 @click.option(
-    "--n_iter",
+    "--prior",
+    default=APGDPriors.NONNEG,
+    type=click.Choice(APGDPriors.all_values()),
+    help="Prior/penalty for APGD.",
+)
+@click.option(
+    "--max_iter",
     type=int,
-    default=5,
-    help="Number of iterations.",
+    default=300,
+    help="Maximum number of iterations.",
 )
 @click.option(
     "--downsample",
@@ -43,8 +61,15 @@ from lensless import ADMM
     help="Downsampling factor.",
 )
 @click.option(
+    "--shape",
+    default=None,
+    nargs=2,
+    type=int,
+    help="Image shape (width, height) for reconstruction.",
+)
+@click.option(
     "--disp",
-    default=1,
+    default=50,
     type=int,
     help="How many iterations to wait for intermediate plot. Set to negative value for no intermediate plots.",
 )
@@ -71,7 +96,7 @@ from lensless import ADMM
 @click.option(
     "--no_plot",
     is_flag=True,
-    help="Whether to no plot.",
+    help="Whether to not plot between iterations.",
 )
 @click.option(
     "--bg",
@@ -94,14 +119,20 @@ from lensless import ADMM
     is_flag=True,
     help="Same PSF for all channels (sum) or unique PSF for RGB.",
 )
-def admm(
+@click.option(
+    "--real_conv",
+    is_flag=True,
+    help="Whether to use real convolution linear operator.",
+)
+def apgd(
     psf_fp,
     data_fp,
-    n_iter,
+    prior,
+    gray,
+    max_iter,
     downsample,
     disp,
     flip,
-    gray,
     bayer,
     bg,
     rg,
@@ -109,7 +140,11 @@ def admm(
     save,
     no_plot,
     single_psf,
+    real_conv,
+    shape
 ):
+
+    plot = not no_plot
     psf, data = load_data(
         psf_fp=psf_fp,
         data_fp=data_fp,
@@ -117,29 +152,35 @@ def admm(
[…]
