### 2022-04-25 56617a4
Remove unused command line arg.
