+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="lensless",
-    version="0.0.1",
+    version="1.0.0",
     author="Eric Bezzam",
     author_email="ebezzam@gmail.com",
     description="Package to control and image with a lensless camera running on a Raspberry Pi.",
