+++ CHANGELOG.md
@@ -1,0 +1,21 @@
+# CHANGELOG.md
+
+## Unreleased
+
+#### Added
+
+- Scripts for collecting MNIST.
+- Option to collect grayscale data.
+
+#### Changed
+
+- Restructure example scripts, i.e. subfolder `recon` for reconstructions.
+- Remove heavy installs from setup (e.g. pycsou, lpips, skikit-image).
+
+#### Bugfix
+
+- 
+
+## 1.0.0 - (2022-03-21)
+
+First version!
