### 2022-04-26 9cca44d
Add more info for contributing and changelog.
