+++ CHANGELOG.md
@@ -4,6 +4,20 @@
 
 #### Added
 
+- 
+
+#### Changed
+
+- 
+
+#### Bugfix
+
+- 
+
+## 1.0.1 - (2022-04-26)
+
+#### Added
+
 - Scripts for collecting MNIST.
 - Option to collect grayscale data.
 
@@ -12,9 +26,6 @@
 - Restructure example scripts, i.e. subfolder `recon` for reconstructions.
 - Remove heavy installs from setup (e.g. pycsou, lpips, skikit-image).
 
-#### Bugfix
-
-- 
 
 ## 1.0.0 - (2022-03-21)
 
