+++ CONTRIBUTING.md
@@ -51,15 +51,20 @@ to run all tests.
 8. Push your new branch to your fork.
 9. Open a pull request with the original repository.
 
-## Deploy new version to PyPi
+## Release new version and deploy to PyPi
 
 
-```
-# Create tag and upload
-git tag -a vX.X.X -m "Description."
-git push origin vX.X.X
-
-# Create package and upload to Pypi
-python setup.py sdist
-python -m twine upload  dist/lensless-1.0.1.tar.gz
-```
+1. Edit the `version` field in `setup.py`.
+2. Create new tag.
+    ```
+    git tag -a vX.X.X -m "Description."
+    git push origin vX.X.X
+3. Create package and upload to Pypi (`pip install twine` if not already done).
+    ```
+    python setup.py sdist
+    python -m twine upload  dist/lensless-X.X.X.tar.gz
+    ```
+4. On [GitHub](https://github.com/LCAV/LenslessPiCam/tags) set 
+the new tag clicking "..." and selecting "Create release", at top right
+selecting "Edit tag", and at the bottom pressing "Publish release".
+5. Update `CHANGELOG.md`!
