### 2022-04-26 af6e6ae
Update script path in paper.
