### 2022-04-27 #4259 @ebezzam
Hi @danasolav thank you for agreeing to be handling editor. I've looked over the list and found the following people to be suitable, given their interest and experience with hardware / embedded systems:
- phckopper
- julianstirling
- ixjlyons
- aquilesC

We were wondering if people outside the list could be asked to review? For example, the authors of the [DiffuserCam tutorial](https://github.com/Waller-Lab/DiffuserCam-Tutorial) whose work we compare against.
