### 2022-05-11 #4259 @danasolav
@THUHoloLab, would you be interested in reviewing this work on Lensless Computational Imaging with a Raspberry Pi ([resository](https://github.com/LCAV/LenslessPiCam) and [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04259/joss.04259/10.21105.joss.04259.pdf)), for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

In case you are not familiar with JOSS, the [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss) focuses on evaluation of [the software](https://github.com/LCAV/LenslessPiCam) as well as a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04259/joss.04259/10.21105.joss.04259.pdf). JOSS reviews take place on GitHub, they are open and usually smooth and streamlined. Note we can be flexible regarding time needed for review.
