### 2022-05-31 88d9ed4
Fix possible shape mismatch when using RFFT, add RGB APGD example.
