### 2022-05-31 0d1d5c5
Fix possible shape mismatch when using RFFT, add RGB APGD example.
