### 2022-06-15 #4259 @danasolav
:wave: @jamesfolberth & @Nikunj-Gupta,

would any of you be willing to review this submission for JOSS?

JOSS is a free, open-source, community driven and developer-friendly online journal
(no publisher is seeking to raise revenue from the volunteer labor of researchers!).

The review process at JOSS is unique: it takes place in a GitHub issue, is open,
and author-reviewer-editor conversations are encouraged.

The checklist-driven reviews involve downloading and installing [the software](https://github.com/LCAV/LenslessPiCam), and inspecting the repository and submitted [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04259/joss.04259/10.21105.joss.04259.pdf) for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments
and improve their submission until acceptance (or withdrawal, if they feel unable to
satisfy the review).

If you cannot review, can you recommend someone from your team to help out?

Kind regards,
Dana Solav
