### 2022-08-25 9a209e0
Fix default coefficients for grayscale conversion.
