+++ lensless/util.py
@@ -61,7 +61,7 @@ def rgb2gray(rgb, weights=None):
 
     """
     if weights is None:
-        weights = np.array([0.299, 0.587, 0.144])
+        weights = np.array([0.299, 0.587, 0.114])
     assert len(weights) == 3
     return np.tensordot(rgb, weights, axes=((2,), 0))
 
