### 2022-09-02 #4259 @raolivei13
Hello,
I'm Richard. Apologies for the late comment, when I was initially informed of this, the email has gotten lost or I may have not received it. I can review the paper, that is what is needed correct?
