### 2022-09-02 #4259 @raolivei13
This is the link to the paper correct? [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04259/joss.04259/10.21105.joss.04259.pdf)
