### 2022-09-09 #4747 @editorialbot
[REVIEW]: LenslessPiCam: A Hardware and Software Platform for Lensless Computational Imaging with a Raspberry Pi
**Submitting author:** @ebezzam (<a href="http://orcid.org/0000-0003-4837-5031">Eric Bezzam</a>)
**Repository:** https://github.com/LCAV/LenslessPiCam
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4
