### 2022-09-12 #4747 @danasolav
@vboomi, @raolivei13 the review process takes place here.
Please see the instructions in the thread above (generate your checklists etc.) and in this [link](https://joss.readthedocs.io/en/latest/review_criteria.html).
Thanks!
