### 2022-10-03 5edd646
Expose channel axis parameter for ssim.
