### 2022-10-19 b088ff2
Added support for grayscale PSFs in io.py
