### 2022-10-24 #4747 @raolivei13
Hello,
there are some points where I made a mistake and will fix it. I will review
the checkpoints at some point this week.

best,
Richard

On Mon, Oct 17, 2022 at 2:02 AM Eric Bezzam ***@***.***>
wrote:

> Hi @raolivei13 <https://github.com/raolivei13> thank you for your review.
> I was wondering if you could elaborate on some of the points you haven't
> checked? I've left a comment on each point below.
>
> Perhaps some of these things were not made clear in the paper, which would
> be great to receive your feedback on how we can better present / what we
> should include to fill in the gaps. Thanks!
> Substantial scholarly effort
>
> Why do you think the work the doesn't meet the scope eligibility described
> in the JOSS guidelines?
> Data sharing
>
> We describe in the README
> <https://github.com/LCAV/LenslessPiCam#data-for-examples-> where to get
> the data for our examples.
> Reproducibility
>
> It's true that when it comes to hardware, it takes more of an effort to
> reproduce. To this end, we tried to be as detailed as possible to reproduce
> our camera though Medium posts
> ***@***.***/a-complete-lensless-imaging-tutorial-hardware-software-and-algorithms-8873fa81a660>.
> Otherwise in terms of reconstruction, we provided scripts
> <https://github.com/LCAV/LenslessPiCam/tree/main/scripts> that we hope
> are straightforward to reproduce the results we present in the paper.
> Functionality
>
> Again, as hardware is involved the functionality for measurement may be
> difficult to reproduce. But in terms of reconstruction, we hope the
> following scripts make it straightforward to confirm that side of things:
>
>    - for individual files
>    <https://github.com/LCAV/LenslessPiCam/tree/main/scripts/recon>
>    - for a dataset
>    <https://github.com/LCAV/LenslessPiCam/blob/main/scripts/evaluate_mirflickr_admm.py>
>
> Performance
>
> The "Efficient reconstruction" section describes some of our performance
> claims, which can be reproduced with these scripts:
>
>    - https://github.com/LCAV/LenslessPiCam/blob/main/profile/admm.py
>    -
>    https://github.com/LCAV/LenslessPiCam/blob/main/profile/gradient_descent.py
>
> Automated tests
>
> We provide unit tests in this folder
> <https://github.com/LCAV/LenslessPiCam/tree/main/test> which can be run
> with pytest
>
