### 2022-11-17 56305fc
Squashed commit of the following:
commit 05b593e7a3ee653c9d834698f0e8780203b449a5
Author: julien-sahli <julien.sahli@epfl.ch>
Date:   Thu Nov 17 14:24:56 2022 +0100

    before stash

commit 6d231bf7079a51f6c9a9cf603a0efa3dab9f33c5
Author: julien-sahli <julien.sahli@epfl.ch>
Date:   Thu Nov 10 15:27:53 2022 +0100

    wip

commit f22eca622b8e9a2bf404b2a0cde0a61b723c6598
Author: julien-sahli <julien.sahli@epfl.ch>
Date:   Thu Nov 10 14:51:06 2022 +0100

    wip

commit 54267710b4432adfe95a898462d767e5c42c5d73
Author: julien-sahli <julien.sahli@epfl.ch>
Date:   Thu Nov 10 14:00:01 2022 +0100

  
[…]
