+++ lensless/io.py
@@ -117,6 +117,7 @@ def load_psf(
     nbits_out=None,
     single_psf=False,
     shape=None,
+    use_3d=False,
 ):
     """
     Load and process PSF for analysis or for reconstruction.
@@ -162,40 +163,60 @@ def load_psf(
     Returns
     -------
     psf :py:class:`~numpy.ndarray`
-        2-D array of PSF.
+        4-D array of PSF [depth, width, height, color]
     """
 
     # load image data and extract necessary channels
-    psf = load_image(
-        fp,
-        verbose=verbose,
-        flip=flip,
-        bayer=bayer,
-        blue_gain=blue_gain,
-        red_gain=red_gain,
-        nbits_out=nbits_out,
-    )
+    if use_3d:
+        assert os.path.isfile(fp)
+        psf = np.load(fp)
+    else:
+        psf = load_image(
+            fp,
+            verbose=verbose,
+            flip=flip,
+            bayer=bayer,
+            blue_gain=blue_gain,
+            red_gain=red_gain,
+            nbits_out=nbits_out,
+        )
 
     original_dtype = psf.dtype
     psf = np.array(psf, dtype=dtype)
 
-    # subtract background, assume black edges
+    if use_3d:
+        if len(psf.shape) == 3:
+            grayscale = True
+            psf = psf[:, :, :, np.newaxis]
+        else:
+            grayscale = False
+    else:
+        if len(psf.shape) == 3:
+            grayscale = False
+            psf = psf[np.newaxis, :, :, :]
+        else:
+            grayscale = True
+            psf = psf[np.newaxis, :, :, np.newaxis]
 
+    assert len(psf.shape) == 4
+    # check that all depths of the psf have the same shape.
+    for i in range(len(psf)):
+        assert psf[0].shape == psf[i].shape
+
+    # subtract background, assume black edges
     if bg_pix is None:
-        bg = np.zeros(len(np.shape(psf)))
+        bg = np.zeros(len(np.shape(psf[0])))
 
     else:
-        # grayscale
-        if len(np.shape(psf)) < 3:
-            bg = np.mean(psf[bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1]])
-            psf -= bg
 
-        # rgb
+        if grayscale:
+            bg = np.mean(psf[:, bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], :])
+            psf -= bg
         else:
             bg = []
             for i in range(3):
-                bg_i = np.mean(psf[bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], i])
-                psf[:, :, i] -= bg_i
+                bg_i = np.mean(psf[:, bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], i])
+                psf[:, :, :, i] -= bg_i
                 bg.append(bg_i)
 
         psf = np.clip(psf, a_min=0, a_max=psf.max())
@@ -203,16 +224,19 @@ def load_psf(
 
     # resize
     if shape:
-        psf = resize(psf, shape=shape)
+        psf = np.array([resize(p, shape=shape) for p in psf])
     elif downsample != 1:
-        psf = resize(psf, factor=1 / downsample)
+        psf = np.array([resize(p, factor=1 / downsample) for p in psf])
 
-    if single_psf:
-        if len(psf.shape) == 3:
+    # todo : ideally resize would need to keep dimensions, but it is used in other places of the code so I'd rather not change it for now
+    if len(psf.shape) == 3:
+        psf = np.array(psf[:, :, :, np.newaxis])
+
+        if not grayscale:
[…]

[…]
