+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 
 #### Added
 
--
+- Support of 3D psfs, which have to be provided as .npy files.
+- 3D reconstriction works for gradient descent and ADMM but the regularizer term will probably need to be adapted
+- The code will automatically perform 3D reconstruction when the provided psf is a .npy file, and perform 2D construction otherwise
 
 #### Changed
 
-- 
+- The data of images and psfs are now always stored as (depth, width, height, color) arrays in memory.
+- Each reconstruction algorith was adapted accordingly.
 
 #### Bugfix
 
