+++ lensless/__init__.py
@@ -7,13 +7,6 @@ from .gradient_descent import (
     GradientDescentUpdate,
 )
 
-from .gradient_descent import (
-    GradientDescient,
-    NesterovGradientDescent,
-    FISTA,
-    GradientDescentUpdate,
-)
-
 try:
     from .apgd import APGD, APGDPriors
 
