### 2022-12-13 #4747 @raolivei13
Hello,
I have reviewed some of the missing points. When it comes to "Reproducibility" I was a little hesitant on checking this off, because it can be interpreted in different ways. Reproducibility in terms of the Software? Or Reproducibility in terms of the Hardware? Since this is a paper which is probably getting published in a Software Journal, we are probably looking at how the Software cab be easily re - used for a similar experiment, which involves some sort of image reconstruction with a Lensless Camera. The Hardware on the other hand, seems to go hand in hand with the Software, i.e "there would be no problem here if we didn't consider the use of a lensless camera". The paper illustrates the problem quite well, but If I were to reproduce this experiment, I would be a little lost on how to set up the experiment on the hardware side of things. For example: "How do I carefully remove the lens from the PiCamera?", "How to capture the Impulse Response of the system (i.e PSF) ?". Not sure if this Hardware Reproducibility is important considering the nature of the paper, which is leaning on the software approach. This is all I have to say, I hope this comment was somewhat insightful, but at the end of the day, as I mentioned before this is a Software paper, and maybe the Hardware aspects of the paper might not be too important.

Best,
Richard
