### 2022-12-13 #4747 @danasolav
@raolivei13, thank you for the important comment.
@ebezzam, since this software is inherently linked to certain hardware. I recommend adding to your repository sufficient details on the hardware and experimental setup, such that replication of your setup by new users is straightforward and unambiguous.
