### 2022-12-13 #4747 @ebezzam
Hi @raolivei13, thank you for that comment on reproducing the hardware. I agree that it is a bit ambiguous whether the hardware is also meant to be reproduced. Nonetheless, this is something we did strive to achieve (reproducibility of hardware *and* accessibility of components), and you can find the instructions on building the camera in the blog post that is referenced in the [README](https://github.com/LCAV/LenslessPiCam/blob/f48870eab405694ec762e016e479a470cee1caa4/README.md?plain=1#L29) and the "About" section of the repository. Moreover, I [just added](https://github.com/LCAV/LenslessPiCam/commit/f48870eab405694ec762e016e479a470cee1caa4) another comment in the Setup section. It is also mentioned in Line 80 of the paper. Please let me know if you think there is another way this information could be made clearer.

We opted for a [Medium article](https://go.epfl.ch/lenslesspicam) as we found it to be a much more friendlier/interactive way to present the hardware side of things:
- The first section links to articles that discuss how to set up the Raspberry Pi.
- The second section links to articles on how to remove the lens, and place the diffuser.
- The third section talks about measuring the PSF and raw data.

But if you feel like more of this info should be place in the README, let us know!
