### 2022-12-21 12e6e07
Add files for prehooks. (#23)
* Add files for pre-hooks.

* Update Contributing file.
