### 2022-12-21 7033a13
Add GitHub Action for building package, formating, and unit tests.
