### 2022-12-21 cab6ee1
Update github action and fix formating.
