### 2022-12-21 5d0d47d
Change exception type to continue.
