+++ lensless/__init__.py
@@ -12,5 +12,5 @@ try:
     from .apgd import APGD, APGDPriors
 
     pycsou_available = True
-except PackageNotFoundError:
+except Exception:
     pycsou_available = False
