+++ setup.py
@@ -26,6 +26,7 @@ setuptools.setup(
         "image==1.5.33",
         "picamerax==20.9.1",
         "matplotlib==3.4.2",
+        "rawpy==0.16.0",
     ],
     extra_requires={"dev": ["pudb", "black"]},
 )
