+++ .github/workflows/python.yml
@@ -47,4 +47,5 @@ jobs:
     - name: Test with pytest
       run: |
         pip install -U pytest
+        pip install -r recon_requirements.txt
         pytest
