+++ setup.py
@@ -22,10 +22,10 @@ setuptools.setup(
         "opencv-python>=4.5.1.48",
         "numpy>=1.22",
         "scipy>=1.7.0",
-        "click==8.0.1",
-        "image==1.5.33",
-        "picamerax==20.9.1",
-        "matplotlib==3.4.2",
+        "click>=8.0.1",
+        "image>=1.5.33",
+        "picamerax>=20.9.1",
+        "matplotlib>=3.4.2",
         "rawpy>=0.16.0",
     ],
     extra_requires={"dev": ["pudb", "black"]},
