+++ setup.py
@@ -19,13 +19,14 @@ setuptools.setup(
     ],
     python_requires=">=3.8.1",
     install_requires=[
-        "opencv-python==4.5.1.48",
+        "opencv-python>=4.5.1.48",
         "numpy>=1.22",
-        "scipy==1.7.0",
-        "click==8.0.1",
-        "image==1.5.33",
-        "picamerax==20.9.1",
-        "matplotlib==3.4.2",
+        "scipy>=1.7.0",
+        "click>=8.0.1",
+        "image>=1.5.33",
+        "picamerax>=20.9.1",
+        "matplotlib>=3.4.2",
+        "rawpy>=0.16.0",
     ],
     extra_requires={"dev": ["pudb", "black"]},
 )
