+++ scripts/remote_display.py
@@ -116,7 +116,7 @@ def remote_display(fp, hostname, pad, vshift, brightness, psf, black, hshift, sc
 
     """ processing on remote machine, less issues with copying """
     # copy picture to Raspberry Pi
-    print(f"\nCopying over picture...")
+    print("\nCopying over picture...")
     os.system('scp %s "pi@%s:%s" ' % (fp, hostname, REMOTE_TMP_PATH))
 
     prep_command = f"{REMOTE_PYTHON} {REMOTE_IMAGE_PREP_SCRIPT} --fp {REMOTE_TMP_PATH} \
