### 2022-12-21 25881c3
Add GitHub Action for building package, formating, and unit tests. (#25)
* Add GitHub Action for building package, formating, and unit tests.

* Update github action and fix formating.

* Change exception type to continue.

* Update dependencies.

* Update dependencies and github workflow.

* Update dependencies.

* Update dependencies.

* Update requirements.
