+++ docs/source/evaluation.rst
@@ -1,0 +1,14 @@
+.. automodule:: lensless.metric
+
+    Metrics API
+    -----------
+
+    .. autofunction:: lensless.metric.mse
+
+    .. autofunction:: lensless.metric.psnr
+
+    .. autofunction:: lensless.metric.ssim
+
+    .. autofunction:: lensless.metric.lpips
+
+    .. autofunction:: lensless.metric.extract

[…]
