+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+sphinx==4.0.1
+sphinx_rtd_theme==0.4.3
+docutils==0.16
