### 2022-12-23 149018a
Change ubuntu version for docs build.
