+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ version: 2
 
 # Set the version of Python and other tools you might need
 build:
-  os: ubuntu-21.04
+  os: ubuntu-20.04
   tools:
     python: "3.9"
 
