### 2022-12-23 c39a3bb
Add mock modules in documentation config.
