+++ lensless/metric.py
@@ -25,11 +25,11 @@ The following script can be used to compute the above metrics for a single file,
 
 .. code:: bash
 
-   python scripts/compute_metrics_from_original.py \\
-   --recon data/reconstruction/admm_thumbs_up_rgb.npy \\
-   --original data/original/thumbs_up.png \\
-   --vertical_crop 262 371 \\
-   --horizontal_crop 438 527 \\
+   python scripts/compute_metrics_from_original.py \
+   --recon data/reconstruction/admm_thumbs_up_rgb.npy \
+   --original data/original/thumbs_up.png \
+   --vertical_crop 262 371 \
+   --horizontal_crop 438 527 \
    --rotation -0.5
 
 where:
@@ -67,8 +67,8 @@ possible to set the number of files.
 
 .. code:: bash
 
-   python scripts/evaluate_mirflickr_admm.py \\
-   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \\
+   python scripts/evaluate_mirflickr_admm.py \
+   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \
    --n_files 10 --save
 
 The ``--save`` flag will save a viewable image for each reconstruction.
@@ -78,8 +78,8 @@ reconstruction.
 
 .. code:: bash
 
-   python scripts/apply_admm_single_mirflickr.py \\
-   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \\
+   python scripts/apply_admm_single_mirflickr.py \
+   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \
    --fid 172
 
 """
