+++ .readthedocs.yaml
@@ -21,5 +21,7 @@ sphinx:
 
 # Optionally declare the Python requirements required to build your docs
 python:
+   version: 3.9
    install:
    - requirements: docs/requirements.txt
+
