### 2022-12-23 4f8d712
Add numpy to docs requirements.
