+++ README.rst
@@ -10,7 +10,7 @@ LenslessPiCam
     :align: center
 
 This package provides functionalities to perform imaging with a lensless
-camera. We make use of a low-cost implementation of DiffuserCam, [1]_
+camera. We make use of a low-cost implementation of DiffuserCam [1]_, 
 where we use a piece of tape instead of the lens and the
 `Raspberry Pi HQ camera sensor <https://www.raspberrypi.com/products/raspberry-pi-high-quality-camera>`__
 (the `V2 sensor <https://www.raspberrypi.com/products/camera-module-v2/>`__
@@ -127,7 +127,7 @@ to them for the idea and making tools/code/data available! Below is some of
 the work that has inspired this toolkit:
 
 * `Build your own DiffuserCam tutorial <https://waller-lab.github.io/DiffuserCam/tutorial>`__.
-* `DiffuserCam Lensless MIR Flickr dataset <https://waller-lab.github.io/LenslessLearning/dataset.html>`__. [2]_
+* `DiffuserCam Lensless MIR Flickr dataset <https://waller-lab.github.io/LenslessLearning/dataset.html>`__ [2]_. 
 
 
 
