+++ docs/source/simulation.rst
@@ -1,0 +1,5 @@
+Simulating raw data
+===================
+
+
+Coming soon! For a sneak peak, check out the code for `LenslessClassification <https://github.com/ebezzam/LenslessClassification>`__.
