+++ .pre-commit-config.yaml
@@ -1,3 +1,4 @@
+exclude: docs
 repos:
 -   repo: https://github.com/psf/black
     rev: 22.12.0
