### 2022-12-23 28bb878
Add documentation files and refactor source files for docs. (#26)
* Add documentation files and refactor source files for docs.

* Fix setup.

* Add version file.

* Add readthedoc config.

* Add comment in requirements.

* Change ubuntu version for docs build.

* Fix version importing.

* Fix version importing.

* Import version from conf.py file.

* Add mock modules in documentation config.

* Fix rendering.

* Set docs requirements.

* Fix readthedocs config.

* Fix requirements.

* Update requrements.

* Update requirements.

* Add numpy to docs requirements.

* Add badge.
