+++ format_code.sh
@@ -6,3 +6,4 @@ black scripts/*.py -l 100
 black scripts/recon/*.py -l 100
 black profile/*.py -l 100
 black test/*.py -l 100
+black docs/source/*.py -l 100
