+++ lensless/metric.py
@@ -1,3 +1,89 @@
+# #############################################################################
+# metric.py
+# =========
+# Authors :
+# Eric BEZZAM [ebezzam@gmail.com]
+# #############################################################################
+
+
+r"""
+Evaluation
+==========
+
+After performing reconstruction, we are typically interested in evaluating the
+quality of the reconstruction. To this end, four metrics are made available:
+
+* **Mean squared error (MSE)**: lower is better with a minimum of 0.
+* **Peak signal-to-noise ratio (PSNR)**: higher is better with values given in decibels (dB).
+* **Structural similarity index measure (SSIM)**: higher is better with a maximum of 1.
+* **Learned Perceptual Image Patch Similarity (LPIPS)**: perceptual metrics that used a pre-trained neural network on patches. Lower is better with a minimum of 0. *NB: only for RGB!*
+
+On a single file
+----------------
+
+The following script can be used to compute the above metrics for a single file, given a reference file:
+
+.. code:: bash
+
+   python scripts/compute_metrics_from_original.py \
+   --recon data/reconstruction/admm_thumbs_up_rgb.npy \
+   --original data/original/thumbs_up.png \
+   --vertical_crop 262 371 \
+   --horizontal_crop 438 527 \
+   --rotation -0.5
+
+where:
+
+*  ``recon`` is the path to the reconstructed file;
+*  ``original`` is the path to the reconstructed file;
+*  ``vertical_crop`` specifies the vertical section to crop from the reconstruction;
+*  ``horizontal_crop`` specifies the horizontal section to crop from the reconstruction;
+*  ``rotation`` specifies a rotate in degrees to align the reconstruction with the original.
+
+More information with an example can be found in
+`this Medium article <https://medium.com/@bezzam/image-similarity-metrics-applied-to-diffusercam-21998967af8d>`__.
+
+
+DiffuserCam Lensless Mirflickr Dataset
+--------------------------------------
+
+The `DiffuserCam Lensless Mirflickr Dataset (DLMD) <https://waller-lab.github.io/LenslessLearning/dataset.html>`__
+comes with (lensed, lensless) image pairs, namely an image that is captured
+with a conventional lensed camera and a corresponding image that is captured
+with the diffuser-based camera.
+
+You can run ADMM on DLMD with the following script.
+
+.. code:: bash
+
+   python scripts/evaluate_mirflickr_admm.py --data <FP>
+
+where ``<FP>`` is the path to the dataset.
+
+However, the original dataset is quite large (25000 files, 100 GB). So
+we've prepared `this subset <https://drive.switch.ch/index.php/s/vmAZzryGI8U8rcE>`__
+(200 files, 725 MB) which you can also pass to the script. It is also
+possible to set the number of files.
+
+.. code:: bash
+
+   python scripts/evaluate_mirflickr_admm.py \
+   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \
+   --n_files 10 --save
+
+The ``--save`` flag will save a viewable image for each reconstruction.
+
+You can also apply ADMM on a single image and visualize the iterative
+reconstruction.
+
+.. code:: bash
+
+   python scripts/apply_admm_single_mirflickr.py \
+   --data DiffuserCam_Mirflickr_200_3011302021_11h43_seed11 \
+   --fid 172
+
+"""
+
 import numpy as np
 from skimage.metrics import mean_squared_error, peak_signal_noise_ratio, structural_similarity
 import lpips as lpips_lib
@@ -7,7 +93,7 @@ from lensless.util import resize
 
 
 def mse(true, est, normalize=True):
-    """
+    r"""
[…]
