+++ docs/source/index.rst
@@ -1,0 +1,29 @@
+.. Welcome to LenslessPiCam's documentation!
+.. =========================================
+
+.. **LenslessPiCam** is a Python library for everything lensless
+.. imaging: designing the system, simulation, performing measurements,
+.. image reconstruction, and evaluation.
+
+.. Check out the :doc:`usage` section for further information, including
+.. how to :ref:`installation` the project.
+
+.. .. note::
+
+..    This project is under active development.
+
+
+.. include:: ../../README.rst
+
+Contents
+--------
+
+.. toctree::
+
+   building
+   measurement
+   simulation
+   data
+   reconstruction
+   evaluation
+   utilities
