+++ CHANGELOG.md
@@ -1,44 +1,0 @@
-# CHANGELOG.md
-
-## Unreleased
-
-#### Added
-
-- Documentation files and configuration, using Sphinx.
-
-#### Changed
-
-- README.md to READ.rst for documentation.
-- Shorten README to separate contents in different pages of docs.
-- Fix typo in GradientDescent class name.
-
-#### Bugfix
-
-- Loading grayscale PSFs would cause an dimension error when removing the background pixels
-
-## 1.0.2 - (2022-05-31)
-
-#### Added
-
-- Example of RGB reconstruction with complex-valued FFT: `scripts/recon/apgd_pycsou.py`
-
-#### Bugfix
-
-- Possible shape mismatch when using the real-valued FFT: forward and backward.
-
-## 1.0.1 - (2022-04-26)
-
-#### Added
-
-- Scripts for collecting MNIST.
-- Option to collect grayscale data.
-
-#### Changed
-
-- Restructure example scripts, i.e. subfolder `recon` for reconstructions.
-- Remove heavy installs from setup (e.g. pycsou, lpips, skikit-image).
-
-
-## 1.0.0 - (2022-03-21)
-
-First version!
