### 2022-12-27 ff9a15d
Add contributing and changelog to docs. (#27)
