+++ lensless/autocorr.py
@@ -1,21 +1,0 @@
-import numpy as np
-import warnings
-
-
-def autocorr2d(vals, pad_mode="reflect"):
-    """
-    Compute 2-D autocorrelation of image via the FFT.
-
-    Parameters
-    ----------
-    vals : py:class:`~numpy.ndarray`
-        2-D image.
-    pad_mode : str
-        Desired padding. See NumPy documentation: https://numpy.org/doc/stable/reference/generated/numpy.pad.html
-
-    Return
-    ------
-    autocorr : py:class:`~numpy.ndarray`
-    """
-    warnings.warn("`autocorr2d` to complete!")
-    return np.ones_like(vals)
