+++ lensless/plot.py
@@ -2,12 +2,7 @@ import numpy as np
 import warnings
 import matplotlib.pyplot as plt
 
-from lensless.util import FLOAT_DTYPES, get_max_val, gamma_correction
-
-try:
-    from lensless.autocorr_sol import autocorr2d
-except Exception:
-    from lensless.autocorr import autocorr2d
+from lensless.util import FLOAT_DTYPES, get_max_val, gamma_correction, autocorr2d
 
 
 def plot_image(img, ax=None, gamma=None, normalize=True):
