+++ docs/source/utilities.rst
@@ -17,3 +17,23 @@ Plotting
     :members:
     :undoc-members:
     :show-inheritance:
+
+
+Image processing
+----------------
+
+.. autofunction:: lensless.util.resize
+
+.. autofunction:: lensless.util.bayer2rgb
+
+.. autofunction:: lensless.util.rgb2gray
+
+.. autofunction:: lensless.util.gamma_correction
+
+
+Image analysis
+--------------
+
+.. autofunction:: lensless.util.autocorr2d
+
+.. autofunction:: lensless.util.print_image_info
