### 2022-12-29 e2e1c12
Add hydra support for evaluation.
