+++ configs/evaluate_mirflickr_admm.yaml
@@ -1,0 +1,22 @@
+# Path to dataset
+# Download data from here: https://drive.switch.ch/index.php/s/vmAZzryGI8U8rcE
+# Or full dataset here: https://github.com/Waller-Lab/LenslessLearning
+dataset: DiffuserCam_Mirflickr_200_3011302021_11h43_seed11
+# null to apply to all
+n_files: 100
+
+preprocess:
+  # Same PSF for all channels (sum) or unique PSF for RGB.
+  single_psf: False
+
+# Whether to save reconstructions.
+save: False
+
+admm:
+  # Number of iterations
+  n_iter: 100
+  # Hyperparameters
+  mu1: 1e-6
+  mu2: 1e-5
+  mu3: 4e-5
+  tau: 0.0001
