+++ configs/apply_admm_single_mirflickr.yaml
@@ -1,0 +1,31 @@
+# Path to dataset
+# Download data from here: https://drive.switch.ch/index.php/s/vmAZzryGI8U8rcE
+# Or full dataset here: https://github.com/Waller-Lab/LenslessLearning
+dataset: DiffuserCam_Mirflickr_200_3011302021_11h43_seed11
+# File ID
+fid: 172
+
+
+preprocess:
+  # Same PSF for all channels (sum) or unique PSF for RGB.
+  single_psf: False
+
+
+display:
+  # How many iterations to wait for intermediate plot. 
+  # Set to negative value for no intermediate plots.
+  disp: 10
+  # Gamma factor for plotting.
+  gamma: null
+
+# Whether to save reconstructions.
+save: False
+
+admm:
+  # Number of iterations
+  n_iter: 100
+  # Hyperparameters
+  mu1: 1e-6
+  mu2: 1e-5
+  mu3: 4e-5
+  tau: 0.0001
