+++ scripts/recon/admm.py
@@ -1,17 +1,16 @@
 """
 Apply ADMM reconstruction.
-
 ```
-python scripts/recon/admm.py --psf_fp data/psf/tape_rgb.png \
---data_fp data/raw_data/thumbs_up_rgb.png --n_iter 5
+python scripts/recon/admm.py
 ```
 
 """
 
+import hydra
+from hydra.utils import to_absolute_path
 import os
 import time
 import pathlib as plib
-import click
 import matplotlib.pyplot as plt
 from datetime import datetime
 import numpy as np
@@ -19,139 +18,58 @@ from lensless.io import load_data
 from lensless import ADMM
 
 
-@click.command()
-@click.option(
-    "--psf_fp",
-    type=str,
-    help="File name for recorded PSF.",
-)
-@click.option(
-    "--data_fp",
-    type=str,
-    help="File name for raw measurement data.",
-)
-@click.option(
-    "--n_iter",
-    type=int,
-    default=5,
-    help="Number of iterations.",
-)
-@click.option(
-    "--downsample",
-    type=float,
-    default=4,
-    help="Downsampling factor.",
-)
-@click.option(
-    "--shape",
-    default=None,
-    nargs=2,
-    type=int,
-    help="Image shape (height, width) for reconstruction.",
-)
-@click.option(
-    "--disp",
-    default=1,
-    type=int,
-    help="How many iterations to wait for intermediate plot. Set to negative value for no intermediate plots.",
-)
-@click.option(
-    "--flip",
-    is_flag=True,
-    help="Whether to flip image.",
-)
-@click.option(
-    "--save",
-    is_flag=True,
-    help="Whether to save intermediate and final reconstructions.",
-)
-@click.option(
-    "--gray",
-    is_flag=True,
-    help="Whether to perform construction with grayscale.",
-)
-@click.option(
-    "--bayer",
-    is_flag=True,
-    help="Whether image is raw bayer data.",
-)
-@click.option(
-    "--no_plot",
-    is_flag=True,
-    help="Whether to no plot.",
-)
-@click.option(
-    "--bg",
-    type=float,
-    help="Blue gain.",
-)
-@click.option(
-    "--rg",
-    type=float,
-    help="Red gain.",
-)
-@click.option(
-    "--gamma",
-    default=None,
-    type=float,
-    help="Gamma factor for plotting.",
-)
-@click.option(
-    "--single_psf",
-    is_flag=True,
-    help="Same PSF for all channels (sum) or unique PSF for RGB.",
-)
-def admm(
-    psf_fp,
-    data_fp,
-    n_iter,
-    downsample,
-    disp,
-    flip,
-    gray,
-    bayer,
-    bg,
-    rg,
-    gamma,
-    save,
-    no_plot,
-    single_psf,
-    shape,
-):
+@hydra.main(version_base=None, config_path="../../configs", config_name="admm_thumbs_up")
+def admm(config):
+
     psf, data = load_data(
-        psf_fp=psf_fp,
-        data_fp=data_fp,
-        downsample=downsample,
-        bayer=bayer,
-        blue_gain=bg,
-        red_gain=rg,
-        plot=not no_plot,
-        flip=flip,
-        gamma=gamma,
-        gray=gray,
-        single_psf=single_psf,
-        shape=shape,
+        psf_fp=to_absolute_path(config["files"]["psf"]),
+        data_fp=to_absolute_path(config["files"]["data"]),
+        downsample=config["preprocess"]["downsample"],
+        bayer=config["preprocess"]["bayer"],
+        blue_gain=config["preprocess"]["blue_gain"],
+        red_gain=config["preprocess"]["red_gain"],
+        plot=config["display"]["plot"],
+        flip=config["preprocess"]["flip"],
[…]

[…]
