### 2022-12-29 a7ec6f6
Some fixes after going through Medium articles.
