+++ docs/source/simulation.rst
@@ -1,14 +1,4 @@
 Simulating raw data
 ===================
 
-
 Coming soon! For a sneak peak, check out the code for `LenslessClassification <https://github.com/ebezzam/LenslessClassification>`__.
-
-
-Install waveprop
-explanation of example scripts
-- lightweight, easy to understand
-- Datasetloader version
-    - good for batches (parallel)
-    - on GPU
-    - pytorch support for ML trainin
