### 2023-01-06 c392a49
Add examples for simulating data. (#31)
* First commit of simulation files.

* Add torch simulation examples.
