+++ lensless/metric.py
@@ -119,6 +119,8 @@ def mse(true, est, normalize=True):
 
     """
     if normalize:
+        true = np.array(true, dtype=np.float32)
+        est = np.array(est, dtype=np.float32)
         true /= true.max()
         est /= est.max()
     return mean_squared_error(image0=true, image1=est)
@@ -145,6 +147,8 @@ def psnr(true, est, normalize=True):
 
     """
     if normalize:
+        true = np.array(true, dtype=np.float32)
+        est = np.array(est, dtype=np.float32)
         true /= true.max()
         est /= est.max()
     return peak_signal_noise_ratio(image_true=true, image_test=est)
@@ -175,9 +179,14 @@ def ssim(true, est, normalize=True, channel_axis=2, **kwargs):
 
     """
     if normalize:
+        true = np.array(true, dtype=np.float32)
+        est = np.array(est, dtype=np.float32)
         true /= true.max()
         est /= est.max()
     return structural_similarity(im1=true, im2=est, channel_axis=channel_axis, **kwargs)
+
+
+LPIPS_MIN_DIM = 31
 
 
 def lpips(true, est, normalize=True):
@@ -198,11 +207,17 @@ def lpips(true, est, normalize=True):
 
     Returns
     -------
-    mssim : float
+    lpips : float
         The LPIPS metric.
 
     """
+    if np.min(true.shape[:2]) < LPIPS_MIN_DIM:
+        raise ValueError(
+            f"LPIPS requires images to be at least {LPIPS_MIN_DIM}x{LPIPS_MIN_DIM} pixels."
+        )
     if normalize:
+        true = np.array(true, dtype=np.float32)
+        est = np.array(est, dtype=np.float32)
         true /= true.max()
         est /= est.max()
     loss_fn = lpips_lib.LPIPS(net="alex", verbose=False)
