+++ recon_requirements.txt
@@ -3,3 +3,4 @@ lpips==0.1.4
 pycsou>=1.0.6
 pylops==1.18.0
 scikit-image==0.19.0rc0
+waveprop>=0.0.3     # for simulation
