+++ configs/simulate_dataset.yaml
@@ -1,0 +1,40 @@
+# RGB simulation and reconstruction
+
+files:
+  # Can be downloaded here: https://drive.switch.ch/index.php/s/Q5OdDQMwhucIlt8
+  dataset: data/celeba_mini
+  image_ext: jpg
+  n_files: 10    # null to use all
+  psf: data/psf/tape_rgb.png
+
+simulation:
+  object_height: [0.1, 0.4]   # range for random height, or scalar
+  random_shift: True
+  # these distance parameters are typically fixed for a given PSF
+  scene2mask: 40e-2
+  mask2sensor: 4e-3
+  # see waveprop.devices
+  sensor: "rpi_hq"
+  snr_db: 40
+  # simulate different sensor resolution
+  output_dim: null    # [H, W] or null
+  # Downsampling for PSF
+  downsample: 8
+  grayscale: False
+  # max val in simulated measured (quantized 8 bits)
+  max_val: 230
+
+save: 
+  bool: True
+  dir: "data/celeba_mini_tape"
+
+admm:
+  bool: True
+  n_iter: 10
+  # Recommend to not display, ok for small number of files, otherwise many windows will pop up!
+  disp_iter: null
+  # Hyperparameters
+  mu1: 1e-6
+  mu2: 1e-5
+  mu3: 4e-5
+  tau: 0.0001
