### 2023-01-13 e344c32
Update simulation documentation. (#32)
