+++ paper/paper.md
@@ -29,51 +29,30 @@ bibliography: paper.bib
 
 # Summary
 
-Lensless imaging seeks to replace/remove the lens in a conventional imaging system. The earliest cameras were in fact 
-lensless, relying on 
-long exposure times to form images on the other end of a small aperture in a 
-darkened room/container (*camera obscura*). The introduction of a lens
-allowed for more light throughput and therefore shorter exposure 
-times, while retaining sharp focus. The 
-incorporation of digital sensors
-readily enabled the use of computational imaging techniques to post-process and enhance raw images (e.g. via deblurring, inpainting, denoising, sharpening).
+Lensless imaging seeks to replace or remove the lens in a conventional imaging system. 
+The earliest cameras were in fact lensless, relying on long exposure times to form images on the other end of a small aperture in a darkened room/container (*camera obscura*). 
+The introduction of a lens allowed for more light throughput and therefore shorter exposure times, while retaining sharp focus. 
+The incorporation of digital sensors readily enabled the use of computational imaging techniques to post-process and enhance raw images (e.g. via deblurring, inpainting, denoising, sharpening).
 Recently, imaging scientists have started leveraging computational imaging as an integral part of lensless imaging systems, allowing them to form viewable images from the highly multiplexed raw measurements of lensless cameras (see [@boominathan2022recent] and references therein for a comprehensive treatment of lensless imaging).
 This represents a real paradigm shift in camera system design as there is more flexibility to cater the hardware to the application at hand (e.g. lightweight or flat designs).
 This increased flexibility comes however at the price of a more demanding post-processing of the raw digital recordings and a tighter integration of sensing and computation, often difficult to achieve in practice due to inefficient interactions between the various communities of scientists involved. 
-With `LenslessPiCam`, we provide an easily accessible hardware and software 
-framework to enable researchers, hobbyists, and students to implement and 
-explore practical and computational aspects of lensless imaging. We also provide
-detailed guides and exercises so that `LenslessPiCam` can be used as an educational 
-resource, and point to results from our graduate-level signal processing course.
+With `LenslessPiCam`, we provide an easily accessible hardware and software framework to enable researchers, hobbyists, and students to implement and explore practical and computational aspects of lensless imaging.
 
 # Statement of need
 
-Being at the interface of hardware, software, and algorithm design, the field of
-lensless imaging necessitates a broad array of competences that might deter 
-newcomers to the field. The purpose of `LenslessPiCam` is to provide a
-complete toolkit with cheap, 
-accessible hardware designs and open-source software, that also achieves 
-satisfactory results in order to explore novel ideas for hardware, software, and
-algorithm design.
+Being at the interface of hardware, software, and algorithm design, the field of lensless imaging necessitates a broad array of competences that might deter newcomers to the field.
+The purpose of `LenslessPiCam` is to provide a complete toolkit with cheap, accessible hardware designs and open-source software, to quickly enable the exploration of novel ideas for hardware, software, and algorithm design.
 
-The DiffuserCam tutorial [@diffusercam] served as a great starting point to the present toolkit as it demonstrates that a working lensless camera can be
[…]
