### 2023-01-26 65e662b
Update JOSS paper. (#33)
