+++ paper/paper.bib
@@ -45,6 +45,13 @@
     doi={10.1109/ICASSP.2019.8682923}
 }
 
+@article{bezzam2022privacy,
+  title={Privacy-Enhancing Optical Embeddings for Lensless Classification},
+  author={Bezzam, Eric and Vetterli, Martin and Simeoni, Matthieu},
+  journal={arXiv preprint arXiv:2211.12864},
+  year={2022}
+}
+
 @misc{pycsou,
     author = {M. Simeoni},
     title = {Pycsou},
@@ -122,4 +129,12 @@
     volume={269},
     pages={543--547},
     year={1983}
+}
+
+@Misc{Yadan2019Hydra,
+  author =       {Omry Yadan},
+  title =        {Hydra - A framework for elegantly configuring complex applications},
+  howpublished = {Github},
+  year =         {2019},
+  url =          {https://github.com/facebookresearch/hydra}
 }
