### 2023-01-26 #4747 @ebezzam
I've updated the PDF as there have been developments in the project, most notably:
- Added new simulation framework so users can test the whole pipeline without having to build the camera. More on that in [this blog post](https://medium.com/@bezzam/simulating-camera-measurements-through-wave-optics-with-pytorch-support-faf3fa620789).
- A ReadTheDocs site: https://lensless.readthedocs.io
- Added the implementations for 2D autocorrelation and the real-valued FFT, such that they are no longer left to be completed in the code base. So we have removed the educational resource section from the paper.
