### 2023-03-10 #4747 @siddiquesalman
The repository builds on this popular [DiffuserCam tutorial from Waller Lab](https://waller-lab.github.io/DiffuserCam/tutorial.html) which provides extensive details on how to build a diffuser based lensless camera and also provides the python implementation of a few algorithms for easy testing. The proposed repository is an extension of the same with some changes. Therefore I have concerns about the usefulness of this submission in its current form and have some suggestions that might improve it.

Strengths:

1. A more structured approach to Pi camera based lensless imaging.
2. Simple framework for remote capture of raw bayer data.
3. Faster implementation of existing traditional reconstruction algorithms.

Weaknesses and Suggestions:

1. Reconstruction Quality: The current reconstruction resolution is really small for the example image you have provided. I believe a controlled smaller scene that occupies only a part of the field of view (FoV) was used to avoid heavy cropping of measurement. However, this is not a very useful scenario in the real world where you have a significant amount of light coming from the whole FoV of the camera (that can span 40-45 degrees on each side for your prototype). So, please include results and measurements for larger scenes displayed on the monitor and real world scenes of objects placed in front of the camera that occupy a large chunk of the FoV. How does your algorithm scale for these larger scenes? Please provide FoV analysis in the corresponding Medium article or the document accompanying the submission. This is important to justify the usefulness of the system.
2. Biscarrat et. al have provided the python implementations of ADMM, GD, FISTA algorithms which are pretty convenient to use for people working in lensless imaging including myself. The package provided here simply extends it to RGB images and provides a faster implementation. One way to improve upon existing implementation would be to incorporate GPU implementation through python packages like JAX or PyTorch. For example, this repository has GPU implementation of ADMM for lensless camera in PyTorch. Similar re-implementation can be done for other algorithms. I understand that GPUs are not always available, it is still a useful feature to have as it allows future integration of machine learning into it.
3. How does your reconstruction software work for different PSFs and captures already available [here](https://github.com/siddiquesalman/flatnet/tree/flatnet-gen) and [here](https://waller-lab.github.io/LenslessLearning/dataset.html)?

Overall, I like the idea of having a more structured documentation and software for Pi Camera based lensless imaging. However, the usefulness of the current draft is not fully justified. Above suggestions should be incorporated to do that.
