+++ lensless/gradient_descent.py
@@ -3,6 +3,13 @@ from lensless.recon import ReconstructionAlgorithm
 import inspect
 from scipy import fft
 
+try:
+    import torch
+
+    torch_available = True
+except ImportError:
+    torch_available = False
+
 
 class GradientDescentUpdate:
     """Gradient descent update techniques."""
@@ -36,12 +43,15 @@ def non_neg(xi):
         Non-negative projection of input.
 
     """
-    xi = np.maximum(xi, 0)
-    return xi
+    is_torch = isinstance(xi, torch.Tensor)
+    if is_torch:
+        return torch.maximum(xi, torch.zeros_like(xi))
+    else:
+        return np.maximum(xi, 0)
 
 
 class GradientDescent(ReconstructionAlgorithm):
-    def __init__(self, psf, dtype=np.float32, proj=non_neg, **kwargs):
+    def __init__(self, psf, dtype=None, proj=non_neg, **kwargs):
         """
         Object for applying projected gradient descent.
 
@@ -52,7 +62,7 @@ class GradientDescent(ReconstructionAlgorithm):
             2D (grayscale) or 3D (RGB) data can be provided and the shape will
             be used to determine which reconstruction (and allocate the
             appropriate memory).
-        dtype : float32 or float64
+        dtype : float32 or float64, default is float32
             Data type to use for optimization.
         proj : :py:class:`function`
             Projection function to apply at each iteration. Default is
@@ -67,48 +77,91 @@ class GradientDescent(ReconstructionAlgorithm):
         return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
 
     def _pad(self, v):
-        vpad = np.zeros(self._padded_shape).astype(v.dtype)
+        if self.is_torch:
+            vpad = torch.zeros(size=self._padded_shape, dtype=v.dtype, device=v.device)
+        else:
+            vpad = np.zeros(self._padded_shape).astype(v.dtype)
         vpad[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
         return vpad
 
     def reset(self):
 
-        # initial guess, half intensity image
-        # for online approach could use last reconstruction
-        psf_flat = self._psf.reshape(-1, self._n_channels)
-        pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
-        x = np.ones(self._psf_shape, dtype=self._dtype) * pixel_start
-        self._image_est = self._pad(x)
+        if self.is_torch:
 
-        # spatial frequency response
-        self._H = fft.rfft2(
-            self._pad(self._psf), norm="ortho", axes=(0, 1), s=self._padded_shape[:2]
-        )
-        self._Hadj = np.conj(self._H)
+            # initial guess, half intensity image
+            # for online approach could use last reconstruction
+            psf_flat = torch.reshape(self._psf, (-1, self._n_channels))
+            pixel_start = (torch.max(psf_flat, dim=0)[0] + torch.min(psf_flat, dim=0)[0]) / 2
+            x = torch.ones_like(self._psf) * pixel_start
+            self._image_est = self._pad(x)
 
-        Hadj_flat = self._Hadj.reshape(-1, self._n_channels)
-        H_flat = self._H.reshape(-1, self._n_channels)
+            # spatial frequency response
+            self._H = torch.fft.rfft2(
+                self._pad(self._psf), norm="ortho", dim=(0, 1), s=self._padded_shape[:2]
+            )
[…]
