+++ lensless/recon.py
@@ -135,6 +135,13 @@ import matplotlib.pyplot as plt
 from scipy.fftpack import next_fast_len
 from lensless.plot import plot_image
 
+try:
+    import torch
+
+    torch_available = True
+except ImportError:
+    torch_available = False
+
 
 class ReconstructionAlgorithm(abc.ABC):
     """
@@ -160,7 +167,7 @@ class ReconstructionAlgorithm(abc.ABC):
 
     """
 
-    def __init__(self, psf, dtype=np.float32):
+    def __init__(self, psf, dtype=None):
         """
         Base constructor. Derived constructor may define new state variables
         here and also reset them in `reset`.
@@ -168,7 +175,7 @@ class ReconstructionAlgorithm(abc.ABC):
         Parameters
         ----------
 
-            psf : :py:class:`~numpy.ndarray`
+            psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
                 Point spread function (PSF) that models forward propagation.
                 2D (grayscale) or 3D (RGB) data can be provided and the shape will
                 be used to determine which reconstruction (and allocate the
@@ -176,32 +183,59 @@ class ReconstructionAlgorithm(abc.ABC):
             dtype : float32 or float64
                 Data type to use for optimization.
         """
+        self.is_torch = False
+        if torch_available:
+            self.is_torch = isinstance(psf, torch.Tensor)
 
-        self._is_rgb = True if len(psf.shape) == 3 else False
+        # prepate shapes for reconstruction
+        self._is_rgb = len(psf.shape) == 3
         if self._is_rgb:
             self._psf = psf
             self._n_channels = 3
         else:
-            self._psf = psf[:, :, np.newaxis]
+            self._psf = psf[:, :, None]
             self._n_channels = 1
         self._psf_shape = np.array(self._psf.shape)
 
-        if dtype:
-            self._psf = self._psf.astype(dtype)
-            self._dtype = dtype
+        # set dtype
+        if dtype is None:
+            if self.is_torch:
+                dtype = torch.float32
+            else:
+                dtype = np.float32
+
+        if self.is_torch:
+
+            if dtype:
+                self._psf = self._psf.type(dtype)
+                self._dtype = dtype
+            else:
+                self._dtype = self._psf.dtype
+            if self._dtype == torch.float32 or dtype == "float32":
+                self._complex_dtype = torch.complex64
+            elif self._dtype == torch.float64 or dtype == "float64":
+                self._complex_dtype = torch.complex128
+            else:
+                raise ValueError(f"Unsupported dtype : {self._dtype}")
+
         else:
-            self._dtype = self._psf.dtype
-        if self._dtype == np.float32 or dtype == "float32":
-            self._complex_dtype = np.complex64
-        elif self._dtype == np.float64 or dtype == "float64":
-            self._complex_dtype = np.complex128
-        else:
-            raise ValueError(f"Unsupported dtype : {self._dtype}")
+
+            if dtype:
+                self._psf = self._psf.astype(dtype)
+                self._dtype = dtype
+            else:
+                self._dtype = self._psf.dtype
+            if self._dtype == np.float32 or dtype == "float32":
+                self._complex_dtype = np.complex64
+            elif self._dtype == np.float64 or dtype == "float64":
+                self._complex_dtype = np.complex128
+            else:
+                raise ValueError(f"Unsupported dtype : {self._dtype}")
 
         # cropping / padding indices
         self._padded_shape = 2 * self._psf_shape[:2] - 1
[…]
