+++ lensless/admm.py
@@ -2,6 +2,13 @@ import numpy as np
 from lensless.recon import ReconstructionAlgorithm
 from scipy import fft
 
+try:
+    import torch
+
+    torch_available = True
+except ImportError:
+    torch_available = False
+
 
 class ADMM(ReconstructionAlgorithm):
     """
@@ -17,7 +24,7 @@ class ADMM(ReconstructionAlgorithm):
     def __init__(
         self,
         psf,
-        dtype=np.float32,
+        dtype=None,
         mu1=1e-6,
         mu2=1e-5,
         mu3=4e-5,
@@ -31,13 +38,13 @@ class ADMM(ReconstructionAlgorithm):
 
         Parameters
         ----------
-        psf : :py:class:`~numpy.ndarray`
+        psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
             Point spread function (PSF) that models forward propagation.
             2D (grayscale) or 3D (RGB) data can be provided and the shape will
             be used to determine which reconstruction (and allocate the
             appropriate memory).
         dtype : float32 or float64
-            Data type to use for optimization.
+            Data type to use for optimization. Default is float32.
         mu1 : float
             Step size for updating primal/dual variables.
         mu2 : float
@@ -61,12 +68,13 @@ class ADMM(ReconstructionAlgorithm):
         self._tau = tau
 
         # call reset() to initialize matrices
-        super(ADMM, self).__init__(psf, dtype)
+        super(ADMM, self).__init__(psf, dtype, pad=False, norm="backward")
+        # super(ADMM, self).__init__(psf, dtype, pad=False, norm="ortho")
 
         # set prior
         if psi is None:
             # use already defined Psi and PsiT
-            self._PsiTPsi = finite_diff_gram(self._padded_shape, self._dtype)
+            self._PsiTPsi = finite_diff_gram(self._padded_shape, self._dtype, self.is_torch)
         else:
             assert psi_adj is not None
             assert psi_gram is not None
@@ -78,12 +86,21 @@ class ADMM(ReconstructionAlgorithm):
             self._PsiT = psi_adj
             self._PsiTPsi = psi_gram(self._padded_shape)
 
-        # precompute_R_divmat
-        self._R_divmat = 1.0 / (
-            self._mu1 * (np.abs(np.conj(self._H) * self._H))
-            + self._mu2 * np.abs(self._PsiTPsi)
-            + self._mu3
-        ).astype(self._complex_dtype)
+        # precompute_R_divmat (self._H computed by constructor with reset())
+        if self.is_torch:
+
+            self._PsiTPsi = self._PsiTPsi.to(self._psf.device)
+            self._R_divmat = 1.0 / (
+                self._mu1 * (torch.abs(self._convolver._Hadj * self._convolver._H))
+                + self._mu2 * torch.abs(self._PsiTPsi)
+                + self._mu3
+            ).type(self._complex_dtype)
+        else:
+            self._R_divmat = 1.0 / (
+                self._mu1 * (np.abs(self._convolver._Hadj * self._convolver._H))
+                + self._mu2 * np.abs(self._PsiTPsi)
+                + self._mu3
+            ).astype(self._complex_dtype)
 
     def _Psi(self, x):
         """
@@ -98,62 +115,59 @@ class ADMM(ReconstructionAlgorithm):
         """
         return finite_diff_adj(U)
 
-    def _crop(self, x):
-        return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
-
-    def _pad(self, v):
-        """adjoint of cropping"""
-        vpad = np.zeros(self._padded_shape).astype(v.dtype)
[…]
