+++ lensless/gradient_descent.py
@@ -43,8 +43,7 @@ def non_neg(xi):
         Non-negative projection of input.
 
     """
-    is_torch = isinstance(xi, torch.Tensor)
-    if is_torch:
+    if torch_available and isinstance(xi, torch.Tensor):
         return torch.maximum(xi, torch.zeros_like(xi))
     else:
         return np.maximum(xi, 0)
@@ -57,13 +56,13 @@ class GradientDescent(ReconstructionAlgorithm):
 
         Parameters
         ----------
-        psf : :py:class:`~numpy.ndarray`
+        psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
             Point spread function (PSF) that models forward propagation.
             2D (grayscale) or 3D (RGB) data can be provided and the shape will
             be used to determine which reconstruction (and allocate the
             appropriate memory).
-        dtype : float32 or float64, default is float32
-            Data type to use for optimization.
+        dtype : float32 or float64
+            Data type to use for optimization. Default is float32.
         proj : :py:class:`function`
             Projection function to apply at each iteration. Default is
             non-negative.
@@ -73,38 +72,21 @@ class GradientDescent(ReconstructionAlgorithm):
         assert callable(proj)
         self._proj = proj
 
-    def _crop(self, x):
-        return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
-
-    def _pad(self, v):
-        if self.is_torch:
-            vpad = torch.zeros(size=self._padded_shape, dtype=v.dtype, device=v.device)
-        else:
-            vpad = np.zeros(self._padded_shape).astype(v.dtype)
-        vpad[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
-        return vpad
-
     def reset(self):
 
         if self.is_torch:
 
             # initial guess, half intensity image
             # for online approach could use last reconstruction
-            psf_flat = torch.reshape(self._psf, (-1, self._n_channels))
-            pixel_start = (torch.max(psf_flat, dim=0)[0] + torch.min(psf_flat, dim=0)[0]) / 2
-            x = torch.ones_like(self._psf) * pixel_start
-            self._image_est = self._pad(x)
-
-            # spatial frequency response
-            self._H = torch.fft.rfft2(
-                self._pad(self._psf), norm="ortho", dim=(0, 1), s=self._padded_shape[:2]
-            )
-            self._Hadj = torch.conj(self._H)
-
-            Hadj_flat = torch.reshape(self._Hadj, (-1, self._n_channels))
-            H_flat = torch.reshape(self._H, (-1, self._n_channels))
+            psf_flat = self._psf.reshape(-1, self._n_channels)
+            pixel_start = (
+                torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
+            ) / 2
+            self._image_est = torch.ones_like(self._psf) * pixel_start
 
             # set step size as < 2 / lipschitz
+            Hadj_flat = self._convolver._Hadj.reshape(-1, self._n_channels)
+            H_flat = self._convolver._H.reshape(-1, self._n_channels)
             self._alpha = torch.real(1.8 / torch.max(torch.abs(Hadj_flat * H_flat), axis=0).values)
 
         else:
@@ -113,62 +95,23 @@ class GradientDescent(ReconstructionAlgorithm):
             # for online approach could use last reconstruction
             psf_flat = self._psf.reshape(-1, self._n_channels)
             pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
[…]
