### 2023-03-23 #4747 @ebezzam
Hi @siddiquesalman, thank you for your comments! It’s great to have feedback from someone else working in lensless imaging.

Thank you for pointing out some of the strengths. I would also like to point out:

- Simulation framework with PyTorch support, which avoids the need to build the camera.
- Framework for remote display of data.
- Low-cost system with Raspberry Pi HQ sensor that is a significant improvement from the [DiffuserCam tutorial](https://waller-lab.github.io/DiffuserCam/tutorial.html).

Your comments reflect that some of the presentation could be improved. Our paper is already quite lengthy with regards to other JOSS papers, so if you have suggestions on what is essential to keep/rearrange that would be useful. @danasolav do you have suggestions on this?

Regarding the weaknesses and suggestions:
1. It wasn't our intention to cherry-pick a controlled scene. It was one that allows for a "meaningful" comparison with the DiffuserCam tutorial, as the quality is very poor with this baseline camera (Fig 1). Otherwise, results from student projects have shown wider scenes displayed on a monitor (line 51, [example project](https://infoscience.epfl.ch/record/291501?ln=en), see page 10), and Figure 5 shows a wider scene. For FoV analysis, do you mean something like Fig 3 in [DiffuserCam3D](https://opg.optica.org/optica/fulltext.cfm?uri=optica-5-1-1&id=380297)?
2. I’ve added GPU / PyTorch support with [this PR](https://github.com/LCAV/LenslessPiCam/pull/34). I’ve also updated Table 1 to reflect computation time when using PyTorch (with and without GPU). Thank you for this suggestion, as the availability of PyTorch / GPU significantly speeds up computation time (800x and 2000x for GD and ADMM respectively with PyTorch + GPU!) and allows for machine learning integration. As our main platform is a Raspberry Pi, we didn’t intend to have PyTorch support, so we still make it an optional dependency such that users can install the library and use the `numpy` versions without having to install PyTorch.
3. Figure 5 shows reconstruction with the [LenslessLearning](https://waller-lab.github.io/LenslessLearning/dataset.html) dataset. And we are actually working to integrate FlatNet by June!

Once again, thank you for going over our work and your suggestions.
