### 2023-03-29 d0b9ee6
Clean up RPi setup.
