+++ recon_requirements.txt
@@ -3,4 +3,6 @@ lpips==0.1.4
 pycsou>=1.0.6
 pylops==1.18.0
 scikit-image==0.19.0rc0
+hydra-core
+click>=8.0.1
 waveprop>=0.0.3     # for simulation
