### 2023-03-31 #4747 @siddiquesalman
Thanks for addressing my comments @ebezzam
1. I think including reconstruction of real world scenes, in addition to the monitor displayed ones, will address this point (including the FoV question). Including reconstruction results like Fig S3 in the DiffuserCam3D supplementary material ([link](https://opticapublishing.figshare.com/articles/journal_contribution/3159989_pdf/5640856)) or Fig 12 in the PhlatCam paper ([link](https://ieeexplore.ieee.org/document/9076617)) would do the job. Comparison with baseline camera is not needed. Since the low cost prototype is a contribution, it's good to have an idea about how it works for these real world objects. If there
2. Thanks for including GPU support for this. This improves the algorithmic/software contribution.
3. Thanks for pointing this out.
4. I think for some of the figures (like Figures 1 and 3), the images can be cropped to save some space as the majority of the background has zeros.
