+++ lensless/rfft_convolve.py
@@ -32,14 +32,10 @@ class RealFFTConvolve2D:
             self.is_torch = True
 
         # prepare shapes for reconstruction
-        self._is_rgb = len(psf.shape) == 3
-        if self._is_rgb:
-            self._psf = psf
-            self._n_channels = 3
-        else:
-            self._psf = psf[:, :, None]
-            self._n_channels = 1
-        self._psf_shape = np.array(self._psf.shape)
+        assert len(psf.shape) == 4
+        self._use_3d = psf.shape[0] != 1
+        self._is_rgb = psf.shape[3] == 3
+        assert self._is_rgb or psf.shape[3] == 1
 
         # set dtype
         if dtype is None:
@@ -49,70 +45,69 @@ class RealFFTConvolve2D:
                 dtype = np.float32
         else:
             if self.is_torch:
-                self._psf = self._psf.type(dtype)
+                self._psf = psf.type(dtype)
             else:
-                self._psf = self._psf.astype(dtype)
+                self._psf = psf.astype(dtype)
 
-        # cropping / padding indices
-        self._padded_shape = 2 * self._psf_shape[:2] - 1
+        self._psf_shape = np.array(self._psf.shape)
+
+        # cropping / padding indexes
+        self._padded_shape = 2 * self._psf_shape[1:3] - 1
         self._padded_shape = np.array([next_fast_len(i) for i in self._padded_shape])
-        self._padded_shape = list(np.r_[self._padded_shape, [self._n_channels]])
-        self._start_idx = (self._padded_shape[:2] - self._psf_shape[:2]) // 2
-        self._end_idx = self._start_idx + self._psf_shape[:2]
+        self._padded_shape = np.r_[self._psf_shape[0], self._padded_shape, self._psf_shape[3]]
+        self._start_idx = (self._padded_shape[1:3] - self._psf_shape[1:3]) // 2
+        self._end_idx = self._start_idx + self._psf_shape[1:3]
         self.pad = pad  # Whether necessary to pad provided data
 
         # precompute filter in frequency domain
         if self.is_torch:
             self._H = torch.fft.rfft2(
-                self._pad(self._psf), norm=norm, dim=(0, 1), s=self._padded_shape[:2]
+                self._pad(self._psf), norm=norm, dim=(1, 2), s=self._padded_shape[1:3]
             )
             self._Hadj = torch.conj(self._H)
             self._padded_data = torch.zeros(size=self._padded_shape, dtype=dtype, device=psf.device)
 
         else:
-            self._H = fft.rfft2(self._pad(self._psf), axes=(0, 1), norm=norm)
+            self._H = fft.rfft2(self._pad(self._psf), axes=(1, 2), norm=norm)
             self._Hadj = np.conj(self._H)
             self._padded_data = np.zeros(self._padded_shape).astype(dtype)
 
     def _crop(self, x):
-        return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
+        return x[:, self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
 
     def _pad(self, v):
         if self.is_torch:
+            #Todo : check that it has the same behaviour now that we added dimensions (it should)
             vpad = torch.zeros(size=self._padded_shape, dtype=v.dtype, device=v.device)
         else:
             vpad = np.zeros(self._padded_shape).astype(v.dtype)
-        vpad[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
+        vpad[:, self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
         return vpad
 
[…]
