### 2023-04-19 955bf9a
Add FOV image.
