+++ README.rst
@@ -79,10 +79,10 @@ install the library locally.
    python scripts/recon/admm.py
 
 
-Note (as of 19-04-2023), there may be a problem setting up Pycsou via ``pip`` as 
-the setup uses asterisks with greater than or equal requirements is no longer
-supported. Therefore it (``pycsou>=1.0.6``) has been removed from
-``recon_requirements.txt``.
+Note (as of 19-04-2023), there is a problem setting up Pycsou via ``pip`` as 
+the `setup file <https://github.com/matthieumeo/pycsou/blob/master/requirements.txt>`__
+uses asterisks with greater than or equal requirements, which is no longer supported.
+Therefore ``pycsou>=1.0.6`` has been removed from ``recon_requirements.txt``.
 
 **Raspberry Pi**
 
