+++ lensless/apgd_v2.py
@@ -1,0 +1,278 @@
+from lensless.recon import ReconstructionAlgorithm
+import inspect
+import numpy as np
+from pycsou.func.loss import SquaredL2Loss
+from pycsou.func.penalty import NonNegativeOrthant, SquaredL2Norm, L1Norm
+from pycsou.opt.proxalgs import APGD as APGD_pyc
+from copy import deepcopy
+from pycsou.core.linop import LinearOperator
+from typing import Union, Optional
+from numbers import Number
+from scipy import fft
+from scipy.fftpack import next_fast_len
+
+
+import pycsou.abc as pyca
+import pycsou.operator.func as func
+import pycsou.opt.solver as solver
+import pycsou.opt.stop as stop
+import pycsou.runtime as pycrt
+import pycsou.util as pycu
+import pycsou.util.ptype as pyct
+
+
+class APGDPriors:
+    """
+    Priors (compatible with Pycsou) for APGD.
+
+    See Pycsou documentation for available penalties:
+    https://matthieumeo.github.io/pycsou/html/api/functionals/pycsou.func.penalty.html?highlight=penalty#module-pycsou.func.penalty
+    """
+
+    L2 = "l2"
+    NONNEG = "nonneg"
+    L1 = "l1"
+
+    @staticmethod
+    def all_values():
+        vals = []
+        for i in inspect.getmembers(APGDPriors):
+            # remove private and protected functions, and this function
+            if not i[0].startswith("_") and not callable(i[1]):
+                vals.append(i[1])
+        return vals
+
+
+class RealFFTConvolve2D(pyca.LinOp):
+    def __init__(self, filter: pyct.NDArray, dtype: Optional[type] = None):
+        """
+        Linear operator that performs convolution in Fourier domain, and assumes
+        real-valued signals.
+
+        Parameters
+        ----------
+        filter :py:class:`~numpy.ndarray`
+            2D filter to use. Must be of shape (height, width, channels) even if
+            only one channel.
+        dtype : float32 or float64
+            Data type to use for optimization.
+        """
+
+        assert len(filter.shape) == 3
+        self._filter_shape = np.array(filter.shape)
+        self._n_channels = filter.shape[2]
+
+        # cropping / padding indices
+        self._padded_shape = 2 * self._filter_shape[:2] - 1
+        self._padded_shape = np.array([next_fast_len(i) for i in self._padded_shape])
+        self._padded_shape = np.r_[self._padded_shape, [self._n_channels]]
+        self._start_idx = (self._padded_shape[:2] - self._filter_shape[:2]) // 2
+        self._end_idx = self._start_idx + self._filter_shape[:2]
+
+        # precompute filter in frequency domain
+        self._H = fft.rfft2(self._pad(filter), axes=(0, 1))
+        self._Hadj = np.conj(self._H)
+        self._padded_data = np.zeros(self._padded_shape).astype(dtype)
+
+        shape = (int(np.prod(self._filter_shape)), int(np.prod(self._filter_shape)))
+        super(RealFFTConvolve2D, self).__init__(shape=shape)
+
+    def _crop(self, x):
+        return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
+
+    def _pad(self, v):
+        vpad = np.zeros(self._padded_shape).astype(v.dtype)
+        vpad[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
+        return vpad
+
+    @pycrt.enforce_precision(i="arr")
+    @pycu.vectorize(i="arr")
+    def apply(self, x: pyct.NDArray) -> pyct.NDArray:
+        self._padded_data[
[…]
