### 2023-04-25 8a0049c
Update to Pycsou v2.
