+++ test/test_algos.py
@@ -6,13 +6,11 @@ try:
     from lensless import GradientDescent, NesterovGradientDescent, FISTA, ADMM, APGD
 
     pycsou_available = True
-    algos = [GradientDescent, NesterovGradientDescent, FISTA, ADMM, APGD]
 
 except ImportError:
     from lensless import GradientDescent, NesterovGradientDescent, FISTA, ADMM
 
     pycsou_available = False
-    algos = [GradientDescent, NesterovGradientDescent, FISTA, ADMM]
 
 
 try:
@@ -25,35 +23,132 @@ except ImportError:
 
 psf_fp = "data/psf/tape_rgb.png"
 data_fp = "data/raw_data/thumbs_up_rgb.png"
+dtypes = ["float32", "float64"]
 downsample = 16
 n_iter = 5
 disp = None
 
 
-def test_algo():
-    for algo in algos:
+def test_gradient_descent():
+    for gray in [True, False]:
+        for dtype in dtypes:
+            for torch_data in torch_vals:
+                psf, data = load_data(
+                    psf_fp=psf_fp,
+                    data_fp=data_fp,
+                    downsample=downsample,
+                    plot=False,
+                    gray=gray,
+                    dtype=dtype,
+                    torch=torch_data,
+                )
+                recon = GradientDescent(psf, dtype=dtype)
+                recon.set_data(data)
+                res = recon.apply(n_iter=n_iter, disp_iter=None, plot=False)
+                if gray:
+                    assert len(psf.shape) == 2
+                else:
+                    assert len(psf.shape) == 3
+                assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
+
+
+def test_admm():
+    for gray in [True, False]:
+        for dtype in dtypes:
+            for torch_data in torch_vals:
+                psf, data = load_data(
+                    psf_fp=psf_fp,
+                    data_fp=data_fp,
+                    downsample=downsample,
+                    plot=False,
+                    gray=gray,
+                    dtype=dtype,
+                    torch=torch_data,
+                )
+                recon = ADMM(psf, dtype=dtype)
+                recon.set_data(data)
+                res = recon.apply(n_iter=n_iter, disp_iter=None, plot=False)
+                if gray:
+                    assert len(psf.shape) == 2
+                else:
+                    assert len(psf.shape) == 3
+                assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
+
+
+def test_nesterov_gradient_descent():
+    for gray in [True, False]:
+        for dtype in dtypes:
+            for torch_data in torch_vals:
+                psf, data = load_data(
+                    psf_fp=psf_fp,
+                    data_fp=data_fp,
+                    downsample=downsample,
+                    plot=False,
+                    gray=gray,
+                    dtype=dtype,
+                    torch=torch_data,
+                )
+                recon = NesterovGradientDescent(psf, dtype=dtype)
+                recon.set_data(data)
+                res = recon.apply(n_iter=n_iter, disp_iter=None, plot=False)
+                if gray:
+                    assert len(psf.shape) == 2
+                else:
+                    assert len(psf.shape) == 3
+                assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
+
+
+def test_fista():
+    for gray in [True, False]:
+        for dtype in dtypes:
+            for torch_data in torch_vals:
+                psf, data = load_data(
+                    psf_fp=psf_fp,
+                    data_fp=data_fp,
+                    downsample=downsample,
+                    plot=False,
+                    gray=gray,
+                    dtype=dtype,
+                    torch=torch_data,
+                )
+                recon = FISTA(psf, dtype=dtype)
+                recon.set_data(data)
[…]
