+++ lensless/apgd.py
@@ -1,15 +1,18 @@
 from lensless.recon import ReconstructionAlgorithm
 import inspect
 import numpy as np
-from pycsou.func.loss import SquaredL2Loss
-from pycsou.func.penalty import NonNegativeOrthant, SquaredL2Norm, L1Norm
-from pycsou.opt.proxalgs import APGD as APGD_pyc
-from copy import deepcopy
-from pycsou.core.linop import LinearOperator
-from typing import Union, Optional
-from numbers import Number
+from typing import Optional
 from scipy import fft
 from scipy.fftpack import next_fast_len
+
+
+import pycsou.abc as pyca
+import pycsou.operator.func as func
+import pycsou.opt.solver as solver
+import pycsou.opt.stop as stop
+import pycsou.runtime as pycrt
+import pycsou.util as pycu
+import pycsou.util.ptype as pyct
 
 
 class APGDPriors:
@@ -34,8 +37,8 @@ class APGDPriors:
         return vals
 
 
-class RealFFTConvolve2D(LinearOperator):
-    def __init__(self, filter, dtype: Optional[type] = None):
+class RealFFTConvolve2D(pyca.LinOp):
+    def __init__(self, filter: pyct.NDArray, dtype: Optional[type] = None):
         """
         Linear operator that performs convolution in Fourier domain, and assumes
         real-valued signals.
@@ -66,7 +69,7 @@ class RealFFTConvolve2D(LinearOperator):
         self._padded_data = np.zeros(self._padded_shape).astype(dtype)
 
         shape = (int(np.prod(self._filter_shape)), int(np.prod(self._filter_shape)))
-        super(RealFFTConvolve2D, self).__init__(shape=shape, dtype=dtype)
+        super(RealFFTConvolve2D, self).__init__(shape=shape)
 
     def _crop(self, x):
         return x[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]]
@@ -76,8 +79,9 @@ class RealFFTConvolve2D(LinearOperator):
         vpad[self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]] = v
         return vpad
 
-    def __call__(self, x: Union[Number, np.ndarray]) -> Union[Number, np.ndarray]:
-        # like here: https://github.com/PyLops/pylops/blob/3e7eb22a62ec60e868ccdd03bc4b54806851cb26/pylops/signalprocessing/ConvolveND.py#L103
+    @pycrt.enforce_precision(i="x")
+    @pycu.vectorize(i="x")
+    def apply(self, x: pyct.NDArray) -> pyct.NDArray:
         self._padded_data[
             self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]
         ] = np.reshape(x, self._filter_shape)
@@ -89,7 +93,9 @@ class RealFFTConvolve2D(LinearOperator):
         )
         return y.ravel()
 
-    def adjoint(self, y: Union[Number, np.ndarray]) -> Union[Number, np.ndarray]:
+    @pycrt.enforce_precision(i="y")
+    @pycu.vectorize(i="y")
+    def adjoint(self, y: pyct.NDArray) -> pyct.NDArray:
         self._padded_data[
             self._start_idx[0] : self._end_idx[0], self._start_idx[1] : self._end_idx[1]
         ] = np.reshape(y, self._filter_shape)
@@ -113,6 +119,8 @@ class APGD(ReconstructionAlgorithm):
         acceleration="BT",
         diff_lambda=0.001,
         prox_lambda=0.001,
+        disp=100,
+        rel_error=1e-6,
         **kwargs
     ):
         """
[…]
