### 2023-04-25 fb2534b
Fix pycsou setup command.
