+++ recon_requirements.txt
@@ -5,4 +5,4 @@ scikit-image==0.19.0rc0
 hydra-core
 click>=8.0.1
 waveprop>=0.0.3     # for simulation
-pip install git+https://github.com/matthieumeo/pycsou.git@v2-dev
+git+https://github.com/matthieumeo/pycsou.git@v2-dev
