### 2023-04-25 c4ed5da
Add benchmark to changelog
