+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 
 -  Documentation files and configuration, using Sphinx.
 -  Implementations for ``autocorr2d`` and ``RealFFTConvolve2D``.
+-  Benchmarking tool for ReconstructionAlgorithm
 
 Changed
 ~~~~~~~
