### 2023-04-25 d85386b
Add workshow without Pycsou to test more versions of Python.
