+++ README.rst
@@ -79,18 +79,22 @@ install the library locally.
    python scripts/recon/admm.py
 
 
-Note (25-04-2023): Pycsou V2 is installed by ``recon_requirements.txt``. If PyTorch
-is installed, you will need to be sure to have PyTorch 2.0 or higher, as Pycsou V2
-is not compatible with earlier version of PyTorch.
+Note (25-04-2023): Pycsou V2 is installed by ``recon_requirements.txt``. If PyTorch is installed, you will 
+need to be sure to have PyTorch 2.0 or higher, as Pycsou V2 is not compatible with earlier
+version of PyTorch. Moreover, Pycsou requires Python within 
+`[3.9, 3.11) <https://github.com/matthieumeo/pycsou/blob/v2-dev/setup.cfg#L28>`__.
 
 Setup for PyTorch 2.0:
 
 .. code:: bash
 
-   # GPU
+   # default should be PyTorch 2.0
+   pip install torch torchvision torchaudio
+
+   # nightly build for GPU
    pip install numpy --pre torch --force-reinstall --extra-index-url https://download.pytorch.org/whl/nightly/cu117
 
-   # CPU
+   # nightly build for  CPU
    pip install --pre torch --extra-index-url https://download.pytorch.org/whl/nightly/cpu
 
 
