### 2023-04-25 c4a3f80
Fix GitHub workflows for Pycsou v2.
