### 2023-04-25 ca6c704
Fix GitHub workflows for Pycsou v2.
