### 2023-04-25 ed6a1da
Fix GitHub workflows for Pycsou v2.
