+++ .github/workflows/python_no_pycsou.yml
@@ -11,7 +11,7 @@ jobs:
       max-parallel: 12
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.8, 3.9, "3.10", "3.11"]
+        python-version: [3.8, "3.11"]
     steps:
     - uses: actions/checkout@v3
     - name: Checkout submodules
