### 2023-04-25 e522bef
Clean up examples.
