### 2023-04-25 2a6d88a
Clean up examples.
