+++ configs/apgd_thumbs_up.yaml
@@ -1,29 +1,0 @@
-# python scripts/recon/apgd_pycsou.py
-defaults:
-  - defaults_recon
-  - _self_
-
-preprocess:
-  # Downsampling factor along X and Y
-  downsample: 4
-
-display:
-  # How many iterations to wait for intermediate plot. 
-  # Set to negative value for no intermediate plots.
-  disp: 100
-  # Whether to plot results.
-  plot: True
-  # Gamma factor for plotting.
-  gamma: null
-
-apgd:
-  # Maximum number of iterations
-  max_iter: 2000
-  acceleration: True     
-  # Differentiable prior / regularization: l2, null
-  diff_penalty: null
-  diff_lambda: 0.001
-  # Proximal prior / regularization: nonneg, l1, null
-  prox_penalty: nonneg
-  prox_lambda: 0.001
-
