+++ lensless/admm.py
@@ -75,7 +75,7 @@ class ADMM(ReconstructionAlgorithm):
         self._tau = tau
 
         #3D ADMM is not supported yet
-        if len(psf.shape[0]) > 1:
+        if psf.shape[0] > 1:
             raise NotImplementedError("3D ADMM is not supported yet, use gradient descent or APGD instead.")
 
         # call reset() to initialize matrices
@@ -201,11 +201,11 @@ class ADMM(ReconstructionAlgorithm):
         # rk = self._convolver._pad(rk)
 
         if self.is_torch:
-            freq_space_result = self._R_divmat * torch.fft.rfft2(rk, dim=(0, 1))
-            self._image_est = torch.fft.irfft2(freq_space_result, dim=(0, 1))
+            freq_space_result = self._R_divmat * torch.fft.rfft2(rk, dim=(1, 2))
+            self._image_est = torch.fft.irfft2(freq_space_result, dim=(1, 2))
         else:
-            freq_space_result = self._R_divmat * fft.rfft2(rk, axes=(0, 1))
-            self._image_est = fft.irfft2(freq_space_result, axes=(0, 1))
+            freq_space_result = self._R_divmat * fft.rfft2(rk, axes=(1, 2))
+            self._image_est = fft.irfft2(freq_space_result, axes=(1, 2))
 
         # self._image_est = self._convolver._crop(res)
 
@@ -308,6 +308,6 @@ def finite_diff_gram(shape, dtype=None, is_torch=False):
         ] = -1
 
     if is_torch:
-        return torch.fft.rfft2(gram, dim=(0, 1))
+        return torch.fft.rfft2(gram, dim=(1, 2))
     else:
-        return fft.rfft2(gram, axes=(0, 1))
+        return fft.rfft2(gram, axes=(1, 2))
