### 2023-04-25 0b64d6b
Remove code duplication of RFFTConvolve2D from APGD.
