### 2023-04-25 a31dad8
removed 3d data
