### 2023-04-26 7b26d42
Add examples and tests for different APGD priors.
