+++ configs/apgd_l2.yaml
@@ -1,0 +1,12 @@
+# python scripts/recon/apgd_pycsou.py -cn apgd_l2
+defaults:
+  - defaults_recon
+  - _self_
+
+apgd:
+  diff_penalty: l2
+  diff_lambda: 0.0001
+  prox_penalty: null
+  # Lipschitz
+  lipschitz_tight: True
+  lipschitz_tol: 0.1
