### 2023-04-26 7cc7936
Add examples and tests for different APGD priors.
