### 2023-04-26 e17d76e
Update paper for Pycsou v2.
