### 2023-04-27 74796aa
Default to download and check for GPU.
