### 2023-04-28 84df53d
Generalize ParallelDataset.
