+++ lensless/benchmark.py
@@ -1,16 +1,15 @@
-"""
+# #############################################################################
+# metric.py
+# =========
+# Authors :
+# Yohann PERRON
+# Eric BEZZAM [ebezzam@gmail.com]
+# #############################################################################
 
-Download subset from here: https://drive.switch.ch/index.php/s/vmAZzryGI8U8rcE
-Or full dataset here: https://github.com/Waller-Lab/LenslessLearning
-
-To use integrated test function place the downloaded folder inside the data folder.
-"""
 
 import glob
 import os
-import pathlib as plib
-from datetime import datetime
-from lensless.io import load_psf
+from lensless.io import load_psf, load_image
 from lensless.util import resize
 import numpy as np
 from tqdm import tqdm
@@ -28,6 +27,9 @@ except ImportError:
 class ParallelDataset(Dataset):
     """
     Dataset consisting of lensless and corresponding lensed image.
+
+    It can be used with a PyTorch DataLoader to load a batch of lensless and corresponding lensed images.
+
     """
 
     def __init__(
@@ -39,34 +41,55 @@ class ParallelDataset(Dataset):
         flip=False,
         transform_lensless=None,
         transform_lensed=None,
+        lensless_fn="diffuser",
+        lensed_fn="lensed",
+        image_ext="npy",
+        **kwargs,
     ):
         """
+        Dataset consisting of lensless and corresponding lensed image. Default parameters are for the DiffuserCam 
+        Lensless Mirflickr Dataset (DLMD).
 
         Parameters
         ----------
-        root_dir : str
-            Path to the test dataset.
-            It is expected to contained a file psf.tiff and two folder :
-              diffuser and lensed containing each pair of test image (lensless and lensed) with the same name as a .npy file.
-        n_files : int or None, optional
-            Metrics will be computed only on the first n_files images.
-            If None, all images are used, by default False
-        background : torch.Tensor or None, optional
-            If not None, background is removed from lensless images, by default None
-        transform_lensless : torch.Transform or None, optional
-            Transform to apply to the lensless images, by default None
-        transform_lensed : torch.Transform or None, optional
-            Transform to apply to the lensed images, by default None
+
+            root_dir : str
+                Path to the test dataset. It is expected to contain two folders: ones of lensless images and one of lensed images.
+            n_files : int or None, optional
+                Metrics will be computed only on the first ``n_files`` images. If None, all images are used, by default False
+            background : :py:class:`~torch.Tensor` or None, optional
+                If not ``None``, background is removed from lensless images, by default ``None``.
+            downsample : int, optional
+                Downsample factor of the lensless images, by default 4.
+            flip : bool, optional
+                If ``True``, lensless images are flipped, by default ``False``.
+            transform_lensless : PyTorch Transform or None, optional
+                Transform to apply to the lensless images, by default None
+            transform_lensed : PyTorch Transform or None, optional
+                Transform to apply to the lensed images, by default None
+            lensless_fn : str, optional
+                Name of the folder containing the lensless images, by default "diffuser".
+            lensed_fn : str, optional
+                Name of the folder containing the lensed images, by default "lensed".
+            image_ext : str, optional
[…]
