### 2023-04-28 f21e71b
Updates docs with benchmarking.
