+++ docs/source/evaluation.rst
@@ -1,7 +1,11 @@
 .. automodule:: lensless.metric
 
-    Metrics API
-    -----------
+
+    Evaluation API
+    --------------
+
+    Metrics
+    ~~~~~~~
 
     .. autofunction:: lensless.metric.mse
 
@@ -12,3 +16,14 @@
     .. autofunction:: lensless.metric.lpips
 
     .. autofunction:: lensless.metric.extract
+
+
+    Benchmarking
+    ~~~~~~~~~~~~
+
+    .. automodule:: lensless.benchmark
+
+    .. autoclass:: lensless.benchmark.ParallelDataset
+       :special-members: __init__
+
+    .. autofunction:: lensless.benchmark.benchmark
