### 2023-04-28 655e3f9
Updates docs with benchmarking.
