### 2023-04-28 a7967cb
Add max_iter to reconstruction constructor.
