### 2023-04-28 9abd326
Add max_iter to reconstruction constructor.
