+++ docs/source/conf.py
@@ -59,6 +59,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.todo",
     "sphinx.ext.viewcode",
+    
 ]
 
 intersphinx_mapping = {
