### 2023-04-28 4e8c7d1
fix n_iter used before assignment
