### 2023-04-28 2fc0892
Rename max_iter to n_iter.
