### 2023-04-28 b673c06
Rename max_iter to n_iter.
