+++ pull_request_template.md
@@ -2,7 +2,7 @@ Thank you for submitting a pull request (PR)! Before hitting the submit button, 
 
 Please also refer to the doc on [contributing](https://lensless.readthedocs.io/en/latest/contributing.html) for more details. 
 
-- [ ] Have you run the tests by doing `pytest`?
+- [ ] Have you run the tests by running `pytest` from the repository root? See the [documentation](git push --set-upstream origin pr_template) for more details.
 - [ ] Do you follow the code formatting for this project? This can be done by setting up pre-commit hookrs as described in the [documentation](https://lensless.readthedocs.io/en/latest/contributing.html#coding-style).
 - [ ] Is there a unit test for the proposed code modification? If the PR addresses an issue, the test should make sure the issue is fixed.
 - [ ] Are there docstrings? Do they follow the [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) style?
