+++ pull_request_template.md
@@ -7,6 +7,6 @@ Please also refer to the doc on [contributing](https://lensless.readthedocs.io/e
 - [ ] Is there a unit test for the proposed code modification? If the PR addresses an issue, the test should make sure the issue is fixed.
 - [ ] Are there docstrings? Do they follow the [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) style?
 - [ ] Have you checked that the doc builds properly and that any new RST file has been added to the repo? How to do that is covered in the [documentation](https://lensless.readthedocs.io/en/latest/contributing.html#building-documentation).
-- [ ] Did you document the proposed change in the CHANGELOG file? It should go under "Unreleased".
+- [ ] Did you document the proposed change(s) in the CHANGELOG file? It should go under "Unreleased".
 
 Happy PR :smiley:
