### 2023-05-02 8f577e0
Added support for 3D APGD
