+++ lensless/rfft_convolve.py
@@ -37,8 +37,6 @@ class RealFFTConvolve2D:
         self._use_3d = psf.shape[0] != 1
         self._is_rgb = psf.shape[3] == 3
         assert self._is_rgb or psf.shape[3] == 1
-        self._psf_shape = np.array(self._psf.shape) #is it still used ?
-
 
         # set dtype
         if dtype is None:
