### 2023-05-02 d9206ae
Add sensor object.
