+++ lensless/sensor.py
@@ -1,0 +1,239 @@
+# #############################################################################
+# sensor.py
+# =================
+# Authors :
+# Eric BEZZAM [ebezzam@gmail.com]
+# #############################################################################
+
+
+"""
+Sensor
+======
+
+TODO
+- couple with lens / mask / aperture to get virtual image
+
+"""
+
+
+import numpy as np
+from enum import Enum
+
+from lensless.util import rgb2gray, resize
+from lensless.io import load_image
+
+
+class SensorOptions(Enum):
+    RPI_HQ = "rpi_hq"
+    RPI_GS = "rpi_gs"
+    RPI_V2 = "rpi_v2"
+    BASLER_287 = "basler_287"
+    BASLER_548 = "basler_548"
+
+    @staticmethod
+    def values():
+        return [dev.value for dev in SensorOptions]
+
+
+class SensorParam:
+    PIXEL_SIZE = "pixel_size"
+    RESOLUTION = "resolution"
+    DIAGONAL = "diagonal"
+    COLOR = "color"
+    BIT_DEPTH = "bit_depth"
+
+
+"""
+Note sensors are in landscape orientation.
+"""
+sensor_dict = {
+    # Raspberry Pi HQ Camera Sensor
+    # datasheet: https://www.arducam.com/sony/imx477/#imx477-datasheet
+    # IMX 477
+    SensorOptions.RPI_HQ.value: {
+        SensorParam.PIXEL_SIZE: np.array([1.55e-6, 1.55e-6]),
+        SensorParam.RESOLUTION: np.array([3040, 4056]),
+        SensorParam.DIAGONAL: 7.857e-3,
+        SensorParam.COLOR: True,
+        SensorParam.BIT_DEPTH: [8, 12],
+    },
+    # Raspberry Pi Global Shutter Camera
+    # https://www.raspberrypi.com/products/raspberry-pi-global-shutter-camera/
+    # IMX 296
+    SensorOptions.RPI_GS.value: {
+        SensorParam.PIXEL_SIZE: np.array([3.45e-6, 3.45e-6]),
+        SensorParam.RESOLUTION: np.array([1088, 1456]),
+        SensorParam.DIAGONAL: 6.3e-3,
+        SensorParam.COLOR: True,
+        SensorParam.BIT_DEPTH: [8, 12],
+    },
+    # Raspberry Pi Camera Module V2
+    # https://www.raspberrypi.com/documentation/accessories/camera.html#hardware-specification
+    # IMX 219
+    SensorOptions.RPI_V2.value: {
+        SensorParam.PIXEL_SIZE: np.array([1.12e-6, 1.12e-6]),
+        SensorParam.RESOLUTION: np.array([2464, 3280]),
+        SensorParam.DIAGONAL: 4.6e-3,
+        SensorParam.COLOR: True,
+        SensorParam.BIT_DEPTH: [8],
+    },
+    # Basler daA720-520um
+    # https://www.baslerweb.com/en/products/cameras/area-scan-cameras/dart/daa720-520um-cs-mount/
+    # IMX 287
+    SensorOptions.BASLER_287.value: {
+        SensorParam.PIXEL_SIZE: np.array([6.9e-6, 6.9e-6]),
+        SensorParam.RESOLUTION: np.array([540, 720]),
+        # SensorParam.DIAGONAL: ,
+        SensorParam.COLOR: False,
+        SensorParam.BIT_DEPTH: [8, 12],
+    },
+    # Basler daA2448-70uc
+    # https://www.baslerweb.com/en/products/cameras/area-scan-cameras/dart/daa2448-70uc-cs-mount/
+    # IMX 548
+    SensorOptions.BASLER_548.value: {
[…]
