+++ lensless/__init__.py
@@ -21,6 +21,7 @@ from .gradient_descent import (
     FISTA,
     GradientDescentUpdate,
 )
+from .sensor import VirtualSensor, SensorOptions
 
 try:
     from .apgd import APGD, APGDPriors
