+++ test/test_sensor.py
@@ -1,0 +1,133 @@
+import numpy as np
+import os
+from lensless import VirtualSensor, SensorOptions
+from lensless.io import load_image, save_image
+from lensless.util import rgb2gray
+
+
+def test_sensor_size():
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+        if np.all(sensor.pixel_size * sensor.resolution > sensor.size):
+            raise ValueError(
+                f"Discrepancy in sensor {sensor_name} size. Pixel size * resolution = {sensor.pixel_size * sensor.resolution}, while size = {sensor.size}"
+            )
+
+
+def test_sensor_landscape():
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+        assert (
+            sensor.resolution[0] <= sensor.resolution[1]
+        ), f"Sensor {sensor_name} is not landscape."
+
+
+def test_virtual_image():
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+
+        for bit_depth in sensor.bit_depth:
+            img = sensor.capture()
+            assert np.all(
+                img.shape == sensor.image_shape
+            ), f"Sensor {sensor_name} image shape is {img.shape}, while expected shape is {sensor.image_shape}."
+            assert (
+                img.max() <= 2**bit_depth - 1
+            ), f"Sensor {sensor_name} image max value is {img.max()}, while expected max value is {2**bit_depth - 1}."
+
+
+def test_virtual_image_from_rgb_file():
+    """
+    rgb > rgb
+    rgb > gray
+    """
+    fp = os.path.join(os.path.dirname(__file__), "..", "data", "original", "thumbs_up.png")
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+
+        for bit_depth in sensor.bit_depth:
+            img = sensor.capture(scene=fp, bit_depth=bit_depth)
+            assert np.all(
+                img.shape == sensor.image_shape
+            ), f"Sensor {sensor_name} image shape is {img.shape}, while expected shape is {sensor.image_shape}."
+            assert (
+                img.max() <= 2**bit_depth - 1
+            ), f"Sensor {sensor_name} image max value is {img.max()}, while expected max value is {2**bit_depth - 1}."
+
+
+def test_virtual_image_from_gray_file():
+    """
+    gray > rgb
+    gray > gray
+    """
+    fp = os.path.join(os.path.dirname(__file__), "..", "data", "original", "mnist_3.png")
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+
+        for bit_depth in sensor.bit_depth:
+            img = sensor.capture(scene=fp, bit_depth=bit_depth)
+            assert np.all(
+                img.shape == sensor.image_shape
+            ), f"Sensor {sensor_name} image shape is {img.shape}, while expected shape is {sensor.image_shape}."
+            assert (
+                img.max() <= 2**bit_depth - 1
+            ), f"Sensor {sensor_name} image max value is {img.max()}, while expected max value is {2**bit_depth - 1}."
+
+
+def test_virtual_image_from_rgb_data():
+    fp = os.path.join(os.path.dirname(__file__), "..", "data", "original", "thumbs_up.png")
+
+    # provided normalized float data
+    img = load_image(fp)
+    img = img.astype(np.float32)
+    img /= img.max()
+    for sensor_name in SensorOptions.values():
+        sensor = VirtualSensor.from_name(sensor_name)
+
[…]
