+++ scripts/recon/outputs/2023-04-18/13-59-33/.hydra/config.yaml
@@ -1,34 +1,0 @@
-input:
-  psf: data/psf/tape_rgb.png
-  data: data/raw_data/thumbs_up_rgb.png
-  dtype: float32
-torch: false
-torch_device: cpu
-preprocess:
-  downsample: 4
-  shape: null
-  flip: false
-  bayer: false
-  blue_gain: null
-  red_gain: null
-  single_psf: false
-  gray: false
-display:
-  disp: 50
-  plot: true
-  gamma: null
-save: true
-gradient_descent:
-  n_iter: 300
-  method: fista
-  nesterov:
-    p: 0
-    mu: 0.9
-  fista:
-    tk: 1
-admm:
-  n_iter: 5
-  mu1: 1.0e-06
-  mu2: 1.0e-05
-  mu3: 4.0e-05
-  tau: 0.0001

[…]
