+++ docs/source/conf.py
@@ -28,7 +28,7 @@ MOCK_MODULES = [
     "pycsou.runtime",
     "pycsou.util",
     "pycsou.util.ptype",
-    "PIL"
+    "PIL",
 ]
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = mock.Mock()
@@ -58,7 +58,6 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.todo",
     "sphinx.ext.viewcode",
-    
 ]
 
 intersphinx_mapping = {
