### 2023-05-10 0f4f366
IO : added support of .npz archives for psfs
