### 2023-05-10 0b13110
updated script for npz
