### 2023-05-10 dbf3d4c
updated doc for available 3d recon
