### 2023-05-10 81c3121
Fixed typo in data.rst
