### 2023-05-11 8915529
Change test image.
