### 2023-05-12 f4ba12b
Add option to warm start algorithms.
