+++ lensless/apgd.py
@@ -203,7 +203,8 @@ class APGD(ReconstructionAlgorithm):
         )
 
     def reset(self):
-        self._image_est = np.zeros(self._original_size, dtype=self._dtype)
+        if self._image_est is None:
+            self._image_est = np.zeros(self._original_size, dtype=self._dtype)
 
     def _update(self):
         res = next(self._apgd.steps())
