+++ lensless/recon.py
@@ -114,27 +114,6 @@ passed to the script:
     python scripts/recon/admm.py -cn <CONFIG_FILENAME_WITHOUT_YAML_EXT>
 
 
-3D example
-----------
-
-It is also possible to reconstruct 3D scenes using :py:class:`~lensless.GradientDescent` or :py:class:`~lensless.APGD`. :py:class:`~lensless.ADMM` does not support 3D reconstruction yet.
-This requires to use a 3D PSF as an input in the form of an .npy or .npz file, which actually is a set of 2D PSFs corresponding to the same diffuser sampled with light sources from different depths.
-The input data for 3D reconstructions is still a 2D image, as collected by the camera. The reconstruction will be able to separate which part of the lensless data corresponds to which 2D PSF,
-and therefore to which depth, effectively generating a 3D reconstruction, which will be outputed in the form of an .npy file. A 2D projection on the depth axis is also displayed to the user.
-
-The same scripts for 2D reconstruction can be used for 3D reconstruction, namely ``scripts/recon/gradient_descent.py`` and ``scripts/recon/apgd_pycsou.py``.
-
-3D data is provided in LenslessPiCam, but it is simulated. Real example data can be obtained from `Waller Lab <https://github.com/Waller-Lab/DiffuserCam/tree/master/example_data>`_.
-For both the simulated data and the data from Waller Lab, it is best to set ``downsample=1``:
-
-.. code:: bash
-
-    python scripts/recon/gradient_descent.py \\
-    input.psf="path/to/3D/psf.npy" \\
-    input.data="path/to/lensless/data.tiff" \\
-    preprocess.downsample=1
-
-
 Other approaches
 ----------------
 
@@ -188,7 +167,7 @@ class ReconstructionAlgorithm(abc.ABC):
 
     """
 
-    def __init__(self, psf, dtype=None, pad=True, n_iter=100, initial_est=None, **kwargs):
+    def __init__(self, psf, dtype=None, pad=True, n_iter=100, **kwargs):
         """
         Base constructor. Derived constructor may define new state variables
         here and also reset them in `reset`.
@@ -208,21 +187,22 @@ class ReconstructionAlgorithm(abc.ABC):
             n_iter : int, optional
                 Number of iterations to run algorithm for. Can be overridden in
                 `apply`.
-            initial_est : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`, optional
-                Initial estimate of the image. If not provided, the initial estimate is
-                set to zero or to the mean of the data, depending on the algorithm.
-
         """
         self.is_torch = False
 
         if torch_available:
             self.is_torch = isinstance(psf, torch.Tensor)
 
-        assert len(psf.shape) == 4, "PSF must be 4D: [depth, width, height, channel]."
-        assert psf.shape[3] == 3 or psf.shape[3] == 1, "PSF must either be rgb (3) or grayscale (1)"
-        self._psf = psf
         self._n_iter = n_iter
 
+        # prepare shapes for reconstruction
+        self._is_rgb = len(psf.shape) == 3
+        if self._is_rgb:
+            self._psf = psf
+            self._n_channels = 3
+        else:
+            self._psf = psf[:, :, None]
+            self._n_channels = 1
         self._psf_shape = np.array(self._psf.shape)
 
         # set dtype
@@ -238,6 +218,7 @@ class ReconstructionAlgorithm(abc.ABC):
                 dtype = np.float32 if dtype == "float32" else np.float64
 
         if self.is_torch:
+
             if dtype:
                 self._psf = self._psf.type(dtype)
                 self._dtype = dtype
@@ -251,6 +232,7 @@ class ReconstructionAlgorithm(abc.ABC):
[…]
