+++ lensless/plot.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 from lensless.util import FLOAT_DTYPES, get_max_val, gamma_correction, autocorr2d
 
 
-def plot_image(img, ax=None, gamma=None, normalize=True, axis=0):
+def plot_image(img, ax=None, gamma=None, normalize=True):
     """
     Plot image data.
 
@@ -19,8 +19,6 @@ def plot_image(img, ax=None, gamma=None, normalize=True, axis=0):
         Gamma correction factor to apply for plots. Default is None.
     normalize : bool
         Whether to normalize data to maximum range. Default is True.
-    axis : int
-        For 3D data, the axis on which to project the data
 
     Returns
     -------
@@ -43,33 +41,12 @@ def plot_image(img, ax=None, gamma=None, normalize=True, axis=0):
     if gamma and gamma > 1:
         img_norm = gamma_correction(img_norm, gamma=gamma)
 
-    # full data format : [depth, width, height, color]
-    if len(img.shape) == 4:
-        if img.shape[3] == 3:  # 3d rgb
-            sum_img = np.sum(img_norm, axis=axis)
-            ax.imshow(sum_img)
-
-        else:
-            assert img.shape[3] == 1  # 3d grayscale with color channel extended
-            sum_img = np.sum(img_norm[:, :, :, 0], axis=axis)
-            ax.imshow(sum_img, cmap="gray")
-
-    # data of length 3 means we have to infer whichever depth or color is missing, based on shape.
-    elif len(img.shape) == 3:
-        if img.shape[2] == 3:  # 2D rgb
-            ax.imshow(img_norm)
-
-        elif img.shape[2] == 1:  # 2D grayscale with color channel extended
-            ax.imshow(img_norm[:, :, 0], cmap="gray")
-
-        else:  # 3D grayscale
-            sum_img = np.sum(img_norm, axis=axis)
-            ax.imshow(sum_img, cmap="gray")
-
-    # data of length 2 means we have only width and height
-    elif len(img.shape) == 2:  # 2D grayscale
+    if len(img.shape) == 3 and img.shape[2] == 3:
+        ax.imshow(img_norm)
+    elif len(img.shape) == 3 and img.shape[2] == 1:
+        ax.imshow(img_norm[:, :, 0], cmap="gray")
+    elif len(img.shape) == 2:
         ax.imshow(img_norm, cmap="gray")
-
     else:
         raise ValueError(f"Unexpected data shape : {img_norm.shape}")
 
