### 2023-05-12 d7b743e
3d support (#50)
* added conversion scripts

* wip, changing branch

* WIP

* test

* Updated documentation

* removed 3d data

* removed 3d data

* Added support for 3D APGD

* updated changelog

* updated changelog

* removed venv

* autoformat with black

* updated tests

* updated tests

* resize with torch

* IO : added support of .npz archives for psfs

* updated script for npz

* updated doc for available 3d recon

* Fixed typo in data.rst

* improved io.py

* improved recon.py

* added negative indexes to rfft_convolve

* fixed indentation error in gradient descent inital estimation

*
[…]
