+++ lensless/io.py
@@ -118,6 +118,7 @@ def load_psf(
     nbits_out=None,
     single_psf=False,
     shape=None,
+    use_3d=False,
 ):
     """
     Load and process PSF for analysis or for reconstruction.
@@ -163,57 +164,88 @@ def load_psf(
     Returns
     -------
     psf : :py:class:`~numpy.ndarray`
-        2-D array of PSF.
+        4-D array of PSF.
     """
 
     # load image data and extract necessary channels
-    psf = load_image(
-        fp,
-        verbose=verbose,
-        flip=flip,
-        bayer=bayer,
-        blue_gain=blue_gain,
-        red_gain=red_gain,
-        nbits_out=nbits_out,
-    )
+    if use_3d:
+        assert os.path.isfile(fp)
+        if fp.endswith(".npy"):
+            psf = np.load(fp)
+        elif fp.endswith(".npz"):
+            archive = np.load(fp)
+            if len(archive.files) > 1:
+                print("Warning: more than one array in .npz archive, using first")
+            elif len(archive.files) == 0:
+                raise ValueError("No arrays in .npz archive")
+            psf = np.load(fp)[archive.files[0]]
+        else:
+            raise ValueError("File format not supported")
+    else:
+        psf = load_image(
+            fp,
+            verbose=verbose,
+            flip=flip,
+            bayer=bayer,
+            blue_gain=blue_gain,
+            red_gain=red_gain,
+            nbits_out=nbits_out,
+        )
 
     original_dtype = psf.dtype
     psf = np.array(psf, dtype=dtype)
 
-    # subtract background, assume black edges
+    if use_3d:
+        if len(psf.shape) == 3:
+            grayscale = True
+            psf = psf[:, :, :, np.newaxis]
+        else:
+            assert len(psf.shape) == 4
+            grayscale = False
 
+    else:
+        if len(psf.shape) == 3:
+            grayscale = False
+            psf = psf[np.newaxis, :, :, :]
+        else:
+            assert len(psf.shape) == 2
+            grayscale = True
+            psf = psf[np.newaxis, :, :, np.newaxis]
+
+    # check that all depths of the psf have the same shape.
+    for i in range(len(psf)):
+        assert psf[0].shape == psf[i].shape
+
+    # subtract background, assume black edges
     if bg_pix is None:
         bg = np.zeros(len(np.shape(psf)))
 
     else:
         # grayscale
-        if len(np.shape(psf)) < 3:
-            bg = np.mean(psf[bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1]])
+        if grayscale:
+            bg = np.mean(psf[:, bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], :])
             psf -= bg
 
         # rgb
         else:
             bg = []
-            for i in range(3):
-                bg_i = np.mean(psf[bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], i])
-                psf[:, :, i] -= bg_i
+            for i in range(psf.shape[3]):
+                bg_i = np.mean(psf[:, bg_pix[0] : bg_pix[1], bg_pix[0] : bg_pix[1], i])
+                psf[:, :, :, i] -= bg_i
                 bg.append(bg_i)
 
         psf = np.clip(psf, a_min=0, a_max=psf.max())
         bg = np.array(bg)
 
     # resize
-    if shape:
-        psf = resize(psf, shape=shape)
-    elif downsample != 1:
-        psf = resize(psf, factor=1 / downsample)
+    psf = resize(psf, shape=shape, factor=1 / downsample)
 
     if single_psf:
-        if len(psf.shape) == 3:
+        if not grayscale:
[…]
