+++ scripts/recon/gradient_descent.py
@@ -27,7 +27,6 @@ from lensless import (
 def gradient_descent(
     config,
 ):
-
     psf, data = load_data(
         psf_fp=to_absolute_path(config.input.psf),
         data_fp=to_absolute_path(config.input.data),
