+++ lensless/apgd.py
@@ -206,7 +206,7 @@ class APGD(ReconstructionAlgorithm):
         if self._image_est is None:
             self._image_est = np.zeros(self._original_size, dtype=self._dtype)
 
-    def _update(self):
+    def _update(self, iter):
         res = next(self._apgd.steps())
         self._image_est[:] = res["x"]
 
