+++ scripts/recon/unrolled_gradient_descent.py
@@ -1,0 +1,202 @@
+"""
+Apply gradient descent.
+
+```
+python scripts/recon/gradient_descent.py
+```
+
+"""
+
+import hydra
+from hydra.utils import to_absolute_path, get_original_cwd
+import os
+import numpy as np
+import time
+import pathlib as plib
+from datetime import datetime
+import matplotlib.pyplot as plt
+from lensless.io import load_data
+from lensless.unrolled_fista import unrolled_FISTA
+from waveprop.dataset_util import SimulatedPytorchDataset
+from lensless.util import rgb2gray
+import torch
+from torchvision import transforms, datasets
+from tqdm import tqdm
+
+def simulate_dataset(config, psf):
+
+    # load dataset
+    transforms_list = [transforms.ToTensor()]
+    data_path = os.path.join(get_original_cwd(), "data")
+    if config.simulation.grayscale:
+        transforms_list.append(transforms.Grayscale())
+    transform = transforms.Compose(transforms_list)
+    if config.files.dataset == "mnist":
+        ds = datasets.MNIST(root=data_path, train=True, download=True, transform=transform)
+    elif config.files.dataset == "fashion_mnist":
+        ds = datasets.FashionMNIST(root=data_path, train=True, download=True, transform=transform)
+    elif config.files.dataset == "cifar10":
+        ds = datasets.CIFAR10(root=data_path, train=True, download=True, transform=transform)
+    else:
+        raise NotImplementedError(f"Dataset {config.files.dataset} not implemented.")
+
+    # convert PSF
+    if config.simulation.grayscale:
+        psf = rgb2gray(psf)
+    if not isinstance(psf, torch.Tensor):
+        psf = transforms.ToTensor()(psf)
+    elif psf.shape[-1] == 3:
+        # Waveprop syntetic dataset expect C H W
+        psf = psf.permute(2, 0, 1)
+
+    # batch_size = config.files.batch_size
+    batch_size = config.training.batch_size
+    n_files = config.files.n_files
+    device_conv = config.torch_device
+    target = config.target
+
+    # check if gpu is available
+    if device_conv == "cuda" and torch.cuda.is_available():
+        device_conv = "cuda"
+    else:
+        device_conv = "cpu"
+
+    # create Pytorch dataset and dataloader
+    if n_files is not None:
+        ds = torch.utils.data.Subset(ds, np.arange(n_files))
+    ds_prop = SimulatedPytorchDataset(
+        dataset=ds, psf=psf, device_conv=device_conv, target=target, **config.simulation
+    )
+    ds_loader = torch.utils.data.DataLoader(dataset=ds_prop, batch_size=batch_size, shuffle=True)
+    return ds_loader
+
+
+@hydra.main(version_base=None, config_path="../../configs", config_name="unrolled_recon")
+def gradient_descent(
+    config,
+):
+    if config.torch_device == "cuda" and torch.cuda.is_available():
+        print("Using GPU for training.")
+        device = "cuda"
+    else:
+        print("Using CPU for training.")
+        device = "cpu"
+
+    torch.autograd.set_detect_anomaly(True)
+    psf, data = load_data(
+        psf_fp=to_absolute_path(config.input.psf),
+        data_fp=to_absolute_path(config.input.data),
+        dtype=config.input.dtype,
+        downsample=config.simulation.downsample,
+        bayer=config.preprocess.bayer,
+        blue_gain=config.preprocess.blue_gain,
+        red_gain=config.preprocess.red_gain,
+        plot=config.display.plot,
+        flip=config.preprocess.flip,
+        gamma=config.display.gamma,
+        gray=config.preprocess.gray,
[…]
