### 2023-05-12 7a3b7bd
customable weigth for lpips
