### 2023-05-12 0ce1ac1
Add support for CelebA dataset
