### 2023-05-12 88f23ce
Added TrainableReconstructionAlgorithm class.
