### 2023-05-12 b159937
Add batch processing to unrolled ADMM
