### 2023-05-12 3a776db
FIx faster convolutions
