+++ lensless/rfft_convolve.py
@@ -101,9 +101,9 @@ class RealFFTConvolve2D:
             self._padded_data = self._pad(x)
         else:
             if self.is_torch:
-                self._padded_data = x.type(self.dtype).to(self._psf.device)
+                self._padded_data = x  # .type(self.dtype).to(self._psf.device)
             else:
-                self._padded_data = x.astype(self.dtype)
+                self._padded_data = x  # .astype(self.dtype)
 
         if self.is_torch:
             conv_output = torch.fft.ifftshift(
@@ -131,9 +131,9 @@ class RealFFTConvolve2D:
             self._padded_data = self._pad(y)
         else:
             if self.is_torch:
-                self._padded_data = y.type(self.dtype).to(self._psf.device)
+                self._padded_data = y  # .type(self.dtype).to(self._psf.device)
             else:
-                self._padded_data = y.astype(self.dtype)
+                self._padded_data = y  # .astype(self.dtype)
 
         if self.is_torch:
             deconv_output = torch.fft.ifftshift(
