### 2023-05-12 b9ec4e8
Better training script with benchmark
