+++ lensless/unrolled_admm.py
@@ -1,0 +1,204 @@
+import numpy as np
+from lensless.trainable_recon import TrainableReconstructionAlgorithm
+from lensless.admm import soft_thresh, finite_diff, finite_diff_adj, finite_diff_gram
+from scipy import fft
+
+try:
+    import torch
+
+    torch_available = True
+except ImportError:
+    torch_available = False
+
+
+class unrolled_ADMM(TrainableReconstructionAlgorithm):
+    """
+    Object for applying ADMM (Alternating Direction Method of Multipliers) with
+    a non-negativity constraint and a total variation (TV) prior.
+
+    Paper about ADMM: https://web.stanford.edu/~boyd/papers/pdf/admm_distr_stats.pdf
+
+    Slides about ADMM: https://web.stanford.edu/class/ee364b/lectures/admm_slides.pdf
+
+    """
+
+    def __init__(
+        self,
+        psf,
+        dtype=None,
+        n_iter=5,
+        mu1=1e-6,
+        mu2=1e-5,
+        mu3=4e-5,
+        tau=0.0001,
+        psi=None,
+        psi_adj=None,
+        psi_gram=None,
+        **kwargs
+    ):
+        """
+
+        Parameters
+        ----------
+        psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
+            Point spread function (PSF) that models forward propagation.
+            2D (grayscale) or 3D (RGB) data can be provided and the shape will
+            be used to determine which reconstruction (and allocate the
+            appropriate memory).
+        dtype : float32 or float64
+            Data type to use for optimization. Default is float32.
+        mu1 : float
+            Step size for updating primal/dual variables.
+        mu2 : float
+            Step size for updating primal/dual variables.
+        mu3 : float
+            Step size for updating primal/dual variables.
+        tau : float
+            Weight for L1 norm of `psi` applied to the image estimate.
+        psi : :py:class:`function`, optional
+            Operator to map image to a space that the image is assumed to be
+            sparse in (hence L1 norm). Default is to use total variation (TV)
+            operator.
+        psi_adj : :py:class:`function`
+            Adjoint of `psi`.
+        psi_gram : :py:class:`function`
+            Function to compute gram of `psi`.
+        """
+
+        # call reset() to initialize matrices
+        super(unrolled_ADMM, self).__init__(psf, dtype, pad=False, norm="backward")
+        # super(ADMM, self).__init__(psf, dtype, pad=False, norm="ortho")
+
+        self._mu1 = torch.nn.Parameter(torch.ones(n_iter) * mu1)
+        self._mu2 = torch.nn.Parameter(torch.ones(n_iter) * mu2)
+        self._mu3 = torch.nn.Parameter(torch.ones(n_iter) * mu3)
+        self._tau = torch.nn.Parameter(torch.ones(n_iter) * tau)
+
+        # set prior
+        if psi is None:
+            # use already defined Psi and PsiT
+            self._PsiTPsi = finite_diff_gram(self._padded_shape, self._dtype, self.is_torch)
+        else:
+            assert psi_adj is not None
+            assert psi_gram is not None
+            assert callable(psi)
+            assert callable(psi_adj)
+            assert callable(psi_gram)
+            # overwrite already defined Psi and PsiT
+            self._Psi = psi
+            self._PsiT = psi_adj
+            self._PsiTPsi = psi_gram(self._padded_shape)
+
+        # precompute_R_divmat (self._H computed by constructor with reset())
+        self._PsiTPsi = self._PsiTPsi.to(self._psf.device)
+        self._R_divmat = 1.0 / (
[…]
