### 2023-05-12 175dcef
Change better parameters for training ADMM
