### 2023-05-12 1f3575a
Add metrics traking through epoch
