### 2023-05-12 41bc4a9
Fix same color for all unrolled algorithm
