### 2023-05-12 7ac0352
Add unrolled algorithm to plot
