### 2023-05-12 cabd151
Add benchmarking script for reconstruction
