### 2023-05-12 da16e0a
Add DiffuserCam as training dataset.
