+++ configs/unrolled_recon.yaml
@@ -27,7 +27,7 @@ preprocess:
 display:
   # How many iterations to wait for intermediate plot. 
   # Set to negative value for no intermediate plots.
-  disp: 250
+  disp: 100
   # Whether to plot results.
   plot: True
   # Gamma factor for plotting.
@@ -59,7 +59,7 @@ gradient_descent:
 # Train Dataset
 
 files:
-  dataset: "CelebA"   # "mnist" or "fashion_mnist" or "cifar10"
+  dataset: "DiffuserCam"   # "mnist", "fashion_mnist", "cifar10", "CelebA", "DiffuserCam"
   n_files: null    # null to use all
 
 target: "object_plane"    # "original" or "object_plane" or "label"
