### 2023-05-12 2aaf5de
Fix n_iter in constructor
