### 2023-05-12 fafe50d
Add separate config for each unrolled algorithm
