### 2023-05-12 9a0e022
Change better plotting of unrolled algorithm
