### 2023-05-12 2ff12e0
Add reconstruction error to benchmark
