### 2023-05-12 8aa4274
Change better and faster NAN detection with hook
