### 2023-05-12 69dc7e5
Add gradiant cliping
