### 2023-05-12 827f485
Add Reconstruction Error to training
