+++ configs/benchmark.yaml
@@ -5,4 +5,4 @@ n_files: 200
 #How much should the image be downsampled
 downsample: 8
 #algorithm to benchmark
-algorithms: ["ADMM", "FISTA", "GradientDescent" "NesterovGradientDescent"]
+algorithms: ["ADMM", "FISTA", "GradientDescent", "NesterovGradientDescent"]
