+++ scripts/recon/train_unrolled.py
@@ -1,3 +1,10 @@
+# #############################################################################
+# trained_unrolled.py
+# =================
+# Authors :
+# Yohann PERRON [yohann.perron@gmail.com]
+# #############################################################################
+
 """
 Apply gradient descent.
 
