+++ lensless/unrolled_admm.py
@@ -1,3 +1,10 @@
+# #############################################################################
+# unrolled_admm.py
+# =================
+# Authors :
+# Yohann PERRON [yohann.perron@gmail.com]
+# #############################################################################
+
 import numpy as np
 from lensless.trainable_recon import TrainableReconstructionAlgorithm
 from lensless.admm import soft_thresh, finite_diff, finite_diff_adj, finite_diff_gram
