+++ lensless/__init__.py
@@ -22,6 +22,10 @@ from .gradient_descent import (
     GradientDescentUpdate,
 )
 
+from .trainable_recon import TrainableReconstructionAlgorithm
+from .unrolled_admm import UnrolledADMM
+from .unrolled_fista import UnrolledFISTA
+
 try:
     from .apgd import APGD, APGDPriors
 
