+++ lensless/benchmark.py
@@ -1,5 +1,5 @@
 # #############################################################################
-# metric.py
+# benchmark.py
 # =========
 # Authors :
 # Yohann PERRON
