+++ configs/benchmark.yaml
@@ -1,3 +1,13 @@
+#Config files for benchmark_recon.py
+#Results for unrolled algorithms must be put in the folder "benchmark/trained_results" as a json file
+
+#Hydra config
+hydra:
+  run:
+    dir: "benchmark/${now:%Y-%m-%d}/${now:%H-%M-%S}"
+  job:
+    chdir: True
+
 # numbers of iterations to benchmark
 n_iter_range: [5, 10, 30, 60, 100, 200, 300]
 # number of files to benchmark
@@ -5,4 +15,5 @@ n_files: 200
 #How much should the image be downsampled
 downsample: 8
 #algorithm to benchmark
-algorithms: ["ADMM", "FISTA", "GradientDescent", "NesterovGradientDescent"]
+algorithms: ["NesterovGradientDescent"]
+
