+++ configs/benchmark.yaml
@@ -21,5 +21,5 @@ n_files: 200
 #How much should the image be downsampled
 downsample: 8
 #algorithm to benchmark
-algorithms: ["NesterovGradientDescent"]
+algorithms: ["ADMM", "FISTA", "GradientDescent", "NesterovGradientDescent"]
 
