+++ docs/source/index.rst
@@ -24,6 +24,7 @@ Contents
    building
    measurement
    reconstruction
+   trainable_reconstruction
    evaluation
    utilities
 
