### 2023-05-12 ad1cd0a
Remove batch_call default implementation
