+++ CHANGELOG.rst
@@ -34,6 +34,8 @@ Changed
 -  Updated to Pycsou V2, as ``pip install pycsou`` (Pycsou V1) may not work on some machines.
 -  Data and PSF are now always stored as 4D Data [depth, width, height, color]. Grayscale data has a color axis of length 1 and 2D data has a depth axis of length 1.
 -  Added batch support to RealFFTConvolve2D.
+-  Added batch support to RealFFTConvolve2D.
+-  ``ReconstructionAlgorithm.update`` now take the number of the current iteration to allow for unrolled algorithms.
 
 Bugfix
 ~~~~~~
