+++ configs/unrolled_recon.yaml
@@ -1,10 +1,3 @@
-# #############################################################################
-# unrolled_recon.py
-# =========
-# Authors :
-# Yohann PERRON  [yohann.perron@gmail.com]
-# #############################################################################
-
 hydra:
   job:
     chdir: True    # change to output folder
