### 2023-05-12 bb24dd2
Fix gray support and unify n_iter setting
