+++ CHANGELOG.rst
@@ -33,7 +33,7 @@ Changed
 -  Shorten README to separate contents in different pages of docs.
 -  Fix typo in GradientDescent class name.
 -  Updated to Pycsou V2, as ``pip install pycsou`` (Pycsou V1) may not work on some machines.
--  Data and PSF are now always stored as 4D Data [depth, width, height, color]. Grayscale data has a color axis of length 1 and 2D data has a depth axis of length 1.
+-  PSF are now always stored as 4D Data [depth, width, height, color], Data are stored as [(batch_size), depth, width, height, color] batch_size being facultative. Grayscale data has a color axis of length 1 and 2D data has a depth axis of length 1.
 -  Added batch support to RealFFTConvolve2D.
 -  ``ReconstructionAlgorithm.update`` now take the number of the current iteration to allow for unrolled algorithms.
 -  ``ReconstructionAlgorithm.apply`` now takes a reset parameter (default true) to automaticaly call reset.

[…]
