+++ masks/mask_class/NfPR.py
@@ -1,283 +1,0 @@
-import numpy as np
-import torch
-from numpy.fft import fftshift, fft2, ifftshift, ifft2
-
-
-
-
-
-def ft2(g, delta):
-    """
-    Compute 2D DFT.
-    Inspired by Listing 2.5 of "Numerical Simulation of Optical Wave Propagation with Examples in
-    MATLAB" (2010).
-    Parameters
-    ----------
-    g : :py:class:`~numpy.ndarray`
-        2D input samples.
-    delta : float or list
-        Sampling period along x-dimension and (optionally) y-dimension [m].
-    """
-    if isinstance(delta, float) or isinstance(delta, int):
-        delta = [delta, delta]
-    assert len(delta) == 2
-    fact = delta[0] * delta[1]
-    if torch.is_tensor(g):
-        return torch.fft.fftshift(
-            torch.fft.fft2(
-                # TODO ifftshift of fftshift?
-                torch.fft.fftshift(g * fact)
-            )
-        )
-    else:
-        res = fftshift(fft2(fftshift(g))) * fact
-        if g.dtype == np.float32 or g.dtype == np.complex64:
-            res = res.astype(np.complex64)
-        return res
-
-
-
-def ift2(G, delta_f):
-    """
-    Compute 2D IDFT.
-    Inspired by Listing 2.6 from "Numerical Simulation of Optical Wave Propagation with Examples in
-    MATLAB" (2010).
-    Parameters
-    ----------
-    g : :py:class:`~numpy.ndarray`
-        2D input samples.
-    delta_f : float or list
-        Frequency interval along x-dimension and (optionally) y-dimension [Hz].
-    """
-    if isinstance(delta_f, float) or isinstance(delta_f, int):
-        delta_f = [delta_f, delta_f]
-    assert len(delta_f) == 2
-    fact = G.shape[0] * G.shape[1] * delta_f[0] * delta_f[1]
-    # fact = 1   # TODO : huge difference when we don't apply factor
-    if torch.is_tensor(G):
-        # fact = pt.tensor([fact], dtype=G.dtype)
-        return torch.fft.ifftshift(
-            torch.fft.ifft2(
-                # TODO ifftshift of fftshift?
-                torch.fft.ifftshift(G * fact)
-            )
-        )
-        # * G.shape[0] * G.shape[1] * delta_f[0] * delta_f[1]
-    else:
-        res = ifftshift(ifft2(ifftshift(G * fact)))
-        if G.dtype == np.complex64:
-            res = res.astype(np.complex64)
-        return res
-
-
-
-def sample_points(N, delta, shift=0, pytorch=False):
-    """
-    Return sample points in 2D.
-    Parameters
-    ----------
-    N : int or list
-        Number of sample points
-    delta: int or float or list
-        Sampling period along x-dimension and (optionally) y-dimension [m].
-    shift : int or float or list
-        Shift from optical axis
-    """
-    if isinstance(N, int):
-        N = [N, N]
-    assert len(N) == 2
-    if isinstance(delta, float) or isinstance(delta, int):
-        delta = [delta, delta]
-    assert len(delta) == 2
-    if isinstance(shift, float) or isinstance(shift, int):
-        shift = [shift, shift]
-    assert len(shift) == 2
-    if pytorch:
-        delta = torch.tensor(delta)
-        shift = torch.tensor(shift)
-        x = torch.arange(-N[1] / 2, N[1] / 2) * delta[1] + shift[1]
-        x = torch.unsqueeze(x, 0)
-        y = torch.arange(-N[0] / 2, N[0] / 2) * delta[0] + shift[0]
[…]
