### 2023-05-15 dcabb75
Fix intitialization without dtype
