### 2023-05-16 b633af4
Change refactoring test_algo
