+++ test/test_algos.py
@@ -1,6 +1,6 @@
+import pytest
 from lensless.io import load_data
 
-# classical algorithm
 try:
     import pycsou
     from lensless import GradientDescent, NesterovGradientDescent, FISTA, ADMM, APGD, APGDPriors
@@ -15,103 +15,69 @@ except ImportError:
 
 try:
     import torch
+    from lensless import UnrolledFISTA, UnrolledADMM
 
     torch_is_available = True
-    torch_vals = [True, False]
     torch.autograd.set_detect_anomaly(True)
+    trainable_algos = [UnrolledFISTA, UnrolledADMM]
 except ImportError:
-    torch_vals = [False]
     torch_is_available = False
+    trainable_algos = []
 
 
 psf_fp = "data/psf/tape_rgb.png"
 data_fp = "data/raw_data/thumbs_up_rgb.png"
 dtypes = ["float32", "float64"]
 downsample = 16
-n_iter = 5
+_n_iter = 5
 
 # classical algorithms
+standard_algos = [GradientDescent, NesterovGradientDescent, FISTA, ADMM]
 
 
-def test_gradient_descent():
+@pytest.mark.parametrize("algorithm", standard_algos)
+def test_recon_numpy(algorithm):
     for gray in [True, False]:
         for dtype in dtypes:
-            for torch_data in torch_vals:
-                psf, data = load_data(
-                    psf_fp=psf_fp,
-                    data_fp=data_fp,
-                    downsample=downsample,
-                    plot=False,
-                    gray=gray,
-                    dtype=dtype,
-                    torch=torch_data,
-                )
-                recon = GradientDescent(psf, dtype=dtype)
-                recon.set_data(data)
-                res = recon.apply(n_iter=n_iter, disp_iter=None, plot=False)
-                assert psf.shape[3] == (1 if gray else 3)
-                assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
-                # assert len(res.shape) == 4, f"Got {len(res.shape)}, expected 4"
+            psf, data = load_data(
+                psf_fp=psf_fp,
+                data_fp=data_fp,
+                downsample=downsample,
+                plot=False,
+                gray=gray,
+                dtype=dtype,
+                torch=False,
+            )
+            recon = algorithm(psf, dtype=dtype)
+            recon.set_data(data)
+            res = recon.apply(n_iter=_n_iter, disp_iter=None, plot=False)
+            assert len(psf.shape) == 4
+            assert psf.shape[3] == (1 if gray else 3)
+            assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
 
 
-def test_admm():
+@pytest.mark.parametrize("algorithm", standard_algos + trainable_algos)
+def test_recon_torch(algorithm):
+    if not torch_is_available:
+        return
     for gray in [True, False]:
         for dtype in dtypes:
-            for torch_data in torch_vals:
-                psf, data = load_data(
-                    psf_fp=psf_fp,
-                    data_fp=data_fp,
-                    downsample=downsample,
-                    plot=False,
-                    gray=gray,
-                    dtype=dtype,
-                    torch=torch_data,
-                )
-                recon = ADMM(psf, dtype=dtype)
-                recon.set_data(data)
-                res = recon.apply(n_iter=n_iter, disp_iter=None, plot=False)
-                assert psf.shape[3] == (1 if gray else 3)
-                assert res.dtype == psf.dtype, f"Got {res.dtype}, expected {dtype}"
-
-
-def test_nesterov_gradient_descent():
-    for gray in [True, False]:
-        for dtype in dtypes:
-            for torch_data in torch_vals:
[…]
