+++ test/test_convolver.py
@@ -43,7 +43,17 @@ def test_conv_np():
 def test_conv_torch():
     psf = torch.rand(5, 47, 29, 3)
     convolver = RealFFTConvolve2D(psf, pad=True)
-    data = torch.rand(12, 5, 47, 29, 3)
+    data = torch.rand(12, 1, 47, 29, 3)
     all_convolved = convolver.convolve(data)
-    part_convolved = convolver.convolve(data[:1, :1, :, :, :1])
-    assert torch.equal(all_convolved[:1, :1, :, :, :1], part_convolved[:1, :1, :, :, :1])
+    part_convolved = convolver.convolve(data[:1, :1, :, :, :])
+    torch.testing.assert_close(all_convolved[:1, :1, :, :, :], part_convolved[:1, :1, :, :, :])
+
+
+def test_conv_torch_2():
+    psf = torch.rand(1, 47, 29, 3)
+    convolver = RealFFTConvolve2D(psf, pad=True)
+    data = torch.rand(12, 1, 47, 29, 3)
+    all_convolved = convolver.convolve(data)
+    part_convolved = convolver.convolve(data[:1, :1, :, :, :])
+    print(all_convolved.shape, part_convolved.shape)
+    torch.testing.assert_close(all_convolved[:1, :1, :, :, :], part_convolved[:1, :1, :, :, :])
