+++ lensless/mask.py
@@ -22,7 +22,7 @@ class Mask(abc.ABC):
                  sensor_size_m: tuple, 
                  feature_size: tuple, 
                  distance_sensor: float, 
-                 wavelength: float) -> None:
+                 wavelength=532e-9):
         """
         Parent mask contructor.
         Attributes common to each type of mask.
@@ -33,18 +33,18 @@ class Mask(abc.ABC):
             size of the sensor (px)
         sensor_size_m: tuple (dim=2)
             size of the sensor (m)
-        feature_size: float
+        feature_size: float (tuple ?)
+            TODO, make it work for tuples (cf fresnel_conv)
             size of the feature (m)
         distance_sensor: float
             distance between the mask and the sensor (m)
-        wavelength: float
-            wavelength (m)
+        wavelength: float, optional
+            wavelength to simulate PSF (m)
         """
         
         self.mask = None
         self.psf = None
         self.phase_mask = None
-        self.height_map = None
         self.sensor_size_px = sensor_size_px
         self.sensor_size_m = sensor_size_m
         self.feature_size = feature_size
@@ -75,45 +75,6 @@ class Mask(abc.ABC):
             dz=self.distance_sensor,
             dtype=np.float32
         )
-        pass
-    
-
-    def shape(self):
-        """
-        Shape of the mask.
-        """
-        return self.mask.shape
-    
-
-    def phase_retrieval(self, lambd, d1, dz, n=1.5, n_iter=10):
-        """
-        Iterative phase retrieval algorithm from the PhlatCam article (https://ieeexplore.ieee.org/document/9076617)
-
-        Parameters
-        ----------
-        lambd: float
-            wavelength (m)
-        d1: float
-            sample period on the sensor i.e. pixel size (m)
-        dz: float
-            propagation distance between the mask and the sensor
-        n: float
-            refractive index of the mask substrate
-        n_iter: int
-            number of iterations
-        """
-        M_p = np.sqrt(self.psf)
-        for _ in progressbar.ProgressBar()(range(n_iter)):
-            M_phi = np.exp(1j * np.angle(fresnel_conv(M_p, lambd, d1, -dz, dtype=np.float32)[0]))
-            M_p = np.sqrt(self.psf) * np.exp(1j * fresnel_conv(M_phi, lambd, d1, dz, dtype=np.float32)[0])    
-        phi = np.angle(M_phi)
-
-        self.phase_mask = phi
-        self.height_map = lambd * phi / (2 * np.pi * (n-1))
-        pass
-
-
-
 
 
 class CodedAperture(Mask):
@@ -122,42 +83,34 @@ class CodedAperture(Mask):
     From the FlatCam article https://arxiv.org/abs/1509.00116
     """
 
-    def __init__(self, 
-                 sensor_size_px: tuple, 
-                 sensor_size_m: tuple, 
-                 feature_size: tuple, 
-                 distance_sensor: float,
-                 wavelength: float, 
-                 method: str, 
-                 n_bits: int) -> None:
+    def __init__(self, method='MLS', n_bits=8, **kwargs):
         """
         Coded aperture mask contructor (FlatCam).
 
         Parameters
         ----------
-        sensor_size_px: tuple (dim=2)
-            size of the sensor (px)
-        sensor_size_m: tuple (dim=2)
-            size of the sensor (m)
-        feature_size: float
-            size of the feature (m)
-        distance_sensor: float
-            distance between the mask and the sensor (m)
-        wavelength: float
-            wavelength (m)
         method: str
             pattern generation method (MURA or MLS)
+            default value: MLS
         n_bits: int
[…]
