### 2023-05-17 95c988d
adding forward propagation coherence test
