+++ test/test_masks.py
@@ -1,4 +1,9 @@
-from lensless.mask import CodedAperture, PhaseContour, FresnelZoneAperture, phase_retrieval
+import numpy as np
+import warnings
+warnings.filterwarnings("ignore", category=RuntimeWarning)
+from lensless.mask import CodedAperture, PhaseContour, FresnelZoneAperture
+from lensless.metric import mse, psnr, ssim
+from waveprop.fresnel import fresnel_conv
 
 
 
@@ -28,9 +33,6 @@ def test_flatcam():
 
 
 def test_phlatcam():
-
-    lambd, sensor_size, nb_px, dz = 532e-9, 5e-3, 256, 0.5e-3
-    d1 = sensor_size / nb_px
     
     mask = PhaseContour(noise_period=(8,8),  
                          sensor_size_px=(256,256), 
@@ -39,6 +41,11 @@ def test_phlatcam():
                          distance_sensor=dz, 
                          wavelength=lambd)
     assert mask.mask.shape == (256, 256)
+
+    Mp = np.sqrt(mask.target_psf) * np.exp(1j * np.angle(fresnel_conv(mask.mask, lambd, d1, dz, dtype=np.float32)[0]))
+    assert mse(abs(Mp), np.sqrt(mask.target_psf)) < 0.1
+    assert psnr(abs(Mp), np.sqrt(mask.target_psf)) > 30
+    assert abs(1 - ssim(abs(Mp), np.sqrt(mask.target_psf), channel_axis=None)) < 0.1
 
 
 def test_fza():
