+++ lensless/__init__.py
@@ -22,9 +22,12 @@ from .gradient_descent import (
     GradientDescentUpdate,
 )
 
-from .trainable_recon import TrainableReconstructionAlgorithm
-from .unrolled_admm import UnrolledADMM
-from .unrolled_fista import UnrolledFISTA
+try:
+    from .trainable_recon import TrainableReconstructionAlgorithm
+    from .unrolled_admm import UnrolledADMM
+    from .unrolled_fista import UnrolledFISTA
+except Exception:
+    pass
 
 try:
     from .apgd import APGD, APGDPriors
