+++ lensless/gradient_descent.py
@@ -81,6 +81,7 @@ class GradientDescent(ReconstructionAlgorithm):
                 pixel_start = (
                     torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
                 ) / 2
+                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
                 self._image_est = torch.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
@@ -92,6 +93,7 @@ class GradientDescent(ReconstructionAlgorithm):
             if self._image_est is None:
                 psf_flat = self._psf.reshape(-1, self._psf_shape[3])
                 pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
+                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
                 self._image_est = np.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
