+++ configs/benchmark.yaml
@@ -5,6 +5,7 @@ hydra:
   job:
     chdir: True
 
+device: "cuda"
 # numbers of iterations to benchmark
 n_iter_range: [5, 10, 30, 60, 100, 200, 300]
 # number of files to benchmark
