### 2023-05-23 e69dcc2
New unrolled algorithm for reconstruction (#44)
* Added support for iteration dependant update

* Fixed autograd

* Added TrainableReconstructionAlgorithm class.

* Added unrolled Fista algorithm

* added test for unrolled FISTA

* save target

* fix syntax

* customable weigth for lpips

* Fix error while training with L1loss

* Added support for multiple reconstruction methods

* Add support for CelebA dataset

* Add better support for batch processing

* Fix plotting error

* Better training script with benchmark

* Add metrics traking through epoch

* first implementation

* added unrolled
[…]
