+++ lensless/gradient_descent.py
@@ -1,7 +1,6 @@
 import numpy as np
 from lensless.recon import ReconstructionAlgorithm
 import inspect
-import warnings
 
 try:
     import torch
@@ -50,30 +49,33 @@ def non_neg(xi):
 
 
 class GradientDescent(ReconstructionAlgorithm):
+    """
+    Object for applying projected gradient descent.
+    """
+
     def __init__(self, psf, dtype=None, proj=non_neg, **kwargs):
         """
-        Object for applying projected gradient descent.
 
         Parameters
         ----------
         psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
             Point spread function (PSF) that models forward propagation.
-            2D (grayscale) or 3D (RGB) data can be provided and the shape will
-            be used to determine which reconstruction (and allocate the
-            appropriate memory).
+            Must be of shape (depth, height, width, channels) even if
+            depth = 1 and channels = 1. You can use :py:func:`~lensless.io.load_psf`
+            to load a PSF from a file such that it is in the correct format.
         dtype : float32 or float64
             Data type to use for optimization. Default is float32.
         proj : :py:class:`function`
             Projection function to apply at each iteration. Default is
             non-negative.
         """
+
         assert callable(proj)
         self._proj = proj
         super(GradientDescent, self).__init__(psf, dtype, **kwargs)
 
     def reset(self):
         if self.is_torch:
-
             # initial guess, half intensity image
             # for online approach could use last reconstruction
             if self._image_est is None:
@@ -81,7 +83,8 @@ class GradientDescent(ReconstructionAlgorithm):
                 pixel_start = (
                     torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
                 ) / 2
-                self._image_est = torch.ones_like(self._psf) * pixel_start
+                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+                self._image_est = torch.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
             Hadj_flat = self._convolver._Hadj.reshape(-1, self._psf_shape[3])
@@ -89,11 +92,11 @@ class GradientDescent(ReconstructionAlgorithm):
             self._alpha = torch.real(1.8 / torch.max(torch.abs(Hadj_flat * H_flat), axis=0).values)
 
         else:
-
             if self._image_est is None:
                 psf_flat = self._psf.reshape(-1, self._psf_shape[3])
                 pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
-                self._image_est = np.ones_like(self._psf) * pixel_start
+                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+                self._image_est = np.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
             Hadj_flat = self._convolver._Hadj.reshape(-1, self._psf_shape[3])
@@ -104,12 +107,12 @@ class GradientDescent(ReconstructionAlgorithm):
         diff = self._convolver.convolve(self._image_est) - self._data
         return self._convolver.deconvolve(diff)
 
-    def _update(self):
+    def _update(self, iter):
         self._image_est -= self._alpha * self._grad()
         self._image_est = self._proj(self._image_est)
 
     def _form_image(self):
-        return self._proj(self._image_est).squeeze()
+        return self._proj(self._image_est)
 
 
 class NesterovGradientDescent(GradientDescent):
@@ -117,21 +120,42 @@ class NesterovGradientDescent(GradientDescent):
     Object for applying projected gradient descent with Nesterov momentum for
[…]
