+++ lensless/admm.py
@@ -42,9 +42,9 @@ class ADMM(ReconstructionAlgorithm):
         ----------
         psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
             Point spread function (PSF) that models forward propagation.
-            2D (grayscale) or 3D (RGB) data can be provided and the shape will
-            be used to determine which reconstruction (and allocate the
-            appropriate memory).
+            Must be of shape (depth, height, width, channels) even if
+            depth = 1 and channels = 1. You can use :py:func:`~lensless.io.load_psf`
+            to load a PSF from a file such that it is in the correct format.
         dtype : float32 or float64
             Data type to use for optimization. Default is float32.
         mu1 : float
@@ -64,7 +64,8 @@ class ADMM(ReconstructionAlgorithm):
         psi_gram : :py:class:`function`
             Function to compute gram of `psi`.
         pad : bool
-            Whether to pad the image with zeros before applying the PSF.
+            Whether to pad the image with zeros before applying the PSF. Default
+            is False, as optimized data is already padded.
         norm : str
             Normalization to use for the convolution. Options are "forward",
             "backward", and "ortho". Default is "backward".
@@ -130,7 +131,8 @@ class ADMM(ReconstructionAlgorithm):
         if self.is_torch:
             # TODO initialize without padding
             if self._image_est is None:
-                self._image_est = torch.zeros(self._padded_shape, dtype=self._dtype).to(
+                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+                self._image_est = torch.zeros([1] + self._padded_shape, dtype=self._dtype).to(
                     self._psf.device
                 )
 
@@ -156,10 +158,11 @@ class ADMM(ReconstructionAlgorithm):
             # self._X_divmat = 1.0 / (torch.ones_like(self._psf) + self._mu1)
 
         else:
-            self._X = np.zeros(self._padded_shape, dtype=self._dtype)
-            # self._U = np.zeros(np.r_[self._padded_shape, [2]], dtype=self._dtype)
             if self._image_est is None:
-                self._image_est = np.zeros_like(self._X)
+                self._image_est = np.zeros([1] + self._padded_shape, dtype=self._dtype)
+
+            # self._U = np.zeros(np.r_[self._padded_shape, [2]], dtype=self._dtype)
+            self._X = np.zeros_like(self._image_est)
             self._U = np.zeros_like(self._Psi(self._image_est))
             self._W = np.zeros_like(self._X)
             if self._image_est.max():
@@ -192,6 +195,15 @@ class ADMM(ReconstructionAlgorithm):
             self._xi + self._mu1 * self._forward_out + self._convolver._pad(self._data)
         )
 
+    def _W_update(self):
+        """Non-negativity update"""
+        if self.is_torch:
+            self._W = torch.maximum(
+                self._rho / self._mu3 + self._image_est, torch.zeros_like(self._image_est)
+            )
+        else:
+            self._W = np.maximum(self._rho / self._mu3 + self._image_est, 0)
+
     def _image_update(self):
         rk = (
             (self._mu3 * self._W - self._rho)
@@ -210,15 +222,6 @@ class ADMM(ReconstructionAlgorithm):
 
         # self._image_est = self._convolver._crop(res)
 
-    def _W_update(self):
-        """Non-negativity update"""
-        if self.is_torch:
-            self._W = torch.maximum(
[…]
