### 2023-05-23 1c5ad7a
Fix proj with numpy
(cherry picked from commit ba992127751949739d6865e4ece80e0d922dde68)
