### 2023-05-23 22324e3
fixed normalization bug for 3D plotting and added 3 projections view
