### 2023-05-24 5ec2868
Fix simulation examples for 3D.
