+++ lensless/gradient_descent.py
@@ -77,14 +77,12 @@ class GradientDescent(ReconstructionAlgorithm):
     def reset(self):
         if self.is_torch:
             # initial guess, half intensity image
-            # for online approach could use last reconstruction
-            if self._image_est is None:
-                psf_flat = self._psf.reshape(-1, self._psf_shape[3])
-                pixel_start = (
-                    torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
-                ) / 2
-                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
-                self._image_est = torch.ones_like(self._psf[None, ...]) * pixel_start
+            psf_flat = self._psf.reshape(-1, self._psf_shape[3])
+            pixel_start = (
+                torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
+            ) / 2
+            # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+            self._image_est = torch.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
             Hadj_flat = self._convolver._Hadj.reshape(-1, self._psf_shape[3])
@@ -92,11 +90,10 @@ class GradientDescent(ReconstructionAlgorithm):
             self._alpha = torch.real(1.8 / torch.max(torch.abs(Hadj_flat * H_flat), axis=0).values)
 
         else:
-            if self._image_est is None:
-                psf_flat = self._psf.reshape(-1, self._psf_shape[3])
-                pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
-                # initialize image estimate as [Batch, Depth, Height, Width, Channels]
-                self._image_est = np.ones_like(self._psf[None, ...]) * pixel_start
+            psf_flat = self._psf.reshape(-1, self._psf_shape[3])
+            pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
+            # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+            self._image_est = np.ones_like(self._psf[None, ...]) * pixel_start
 
             # set step size as < 2 / lipschitz
             Hadj_flat = self._convolver._Hadj.reshape(-1, self._psf_shape[3])
