### 2023-05-24 538fe6f
Add function to set initial estimate.
