### 2023-05-24 c3c2521
Add message if reset=False
