### 2023-05-25 e7e3be4
Update docstring.
