### 2023-05-25 59f0349
Fix simulation examples. (#52)
* Fix simulation examples for 3D.

* Fix single file example.

* Add function to set initial estimate.

* Add message if reset=False

* Update CHANGELOG

* Update docstring.
