+++ lensless/gradient_descent.py
@@ -76,9 +76,10 @@ class GradientDescent(ReconstructionAlgorithm):
 
     def reset(self):
         if self.is_torch:
-            # initial guess, half intensity image
-            # for online approach could use last reconstruction
-            if self._image_est is None:
+            if self._initial_est is not None:
+                self._image_est = self._initial_est
+            else:
+                # initial guess, half intensity image
                 psf_flat = self._psf.reshape(-1, self._psf_shape[3])
                 pixel_start = (
                     torch.max(psf_flat, axis=0).values + torch.min(psf_flat, axis=0).values
@@ -92,7 +93,9 @@ class GradientDescent(ReconstructionAlgorithm):
             self._alpha = torch.real(1.8 / torch.max(torch.abs(Hadj_flat * H_flat), axis=0).values)
 
         else:
-            if self._image_est is None:
+            if self._initial_est is not None:
+                self._image_est = self._initial_est
+            else:
                 psf_flat = self._psf.reshape(-1, self._psf_shape[3])
                 pixel_start = (np.max(psf_flat, axis=0) + np.min(psf_flat, axis=0)) / 2
                 # initialize image estimate as [Batch, Depth, Height, Width, Channels]
