### 2023-05-26 427f532
Fix normalization
