### 2023-05-29 #4747 @danasolav
@ebezzam, please see the following minor comments regarding the paper. After addressing these, we'll be able to proceed with the acceptance process:

- [ ] General comment: Please avoid using footnotes for references. For citing existing works/codes, use standard references. It is o.k. to use embedded links for parts of your own repository and e.g., link to the RPi camera version, etc. Footnotes are o.k. for mentioning a side note, but in some cases, it may be preferable to embed them in the main text (see below).
- [ ] Line 15: add parentheses before "see"
- [ ] Line 18: add a comma after e.g.
- [ ] Line 19: add commas before and after however
- [ ] Line 33+39: add the links as embedded links instead of footnotes.
- [ ] Line 40: please add more information on the comparison shown in Figure 1. They don't seem to compare a reconstruction of the same image, so what is the significance of this comparison?
- [ ] Figure 2: please provide more details on the algorithms shown in each row, and avoid using a footnote for links.
- [ ] Line 56: Please correct to: "As opposed to Figure 1 and Figure 2, which show reconstructions of back-illuminated objects..."
- [ ] Line 59: The acronym FOV should appear here after the text.
- [ ] Line 74: Add "provide" at the beginning
- [ ] Line 79: add a comma after G
- [ ] Line 81: correct to "functionalities"
- [ ] Line 86: the acronyms should be defined when they first appear in the text.
- [ ] Line 95: the function name does not compile properly and extends beyond the line. Could you force a manual new line?
- [ ] Line 106: add a comma after i.e.
- [ ] Line 117: refer to Table 1 rather than "the table below"
- [ ] Line 125: refer to Table 1 rather than "the above table"
- [ ] Line 130: add the full definition of FFT
- [ ] Line 132: should be pre-computes?
- [ ] Line 134: add a comma before "respectively".
- [ ] Line 139-142: use ; instead of , at the end of the bullet points, as in the previous bulleted lists.
- [ ] Line 143: use a reference to Figure 6 instead of "Below".
- [ ] Line 143: it would be helpful to explain the meaning of these numbers and their limits, where applicable.
- [ ] Line 143: You have to refer to Table 2 in the text.
- [ ] Line 143-155: use standard references for others' repositories (6,7) and embedded links for yours (4,5,8,9).
- [ ] Line 147-148: should be 25,000
- [ ] Line 165-169: use ; instead of , at the end of the bullet points, as in the previous bulleted lists.

Dana
