### 2023-05-31 ea65341
corrected CodedAperture rows and columns
