+++ lensless/admmPnP.py
@@ -1,0 +1,234 @@
+import numpy as np
+from lensless.recon import ReconstructionAlgorithm
+from scipy import fft
+
+try:
+    import torch
+
+    torch_available = True
+except ImportError:
+    torch_available = False
+
+
+class ADMM_PnP(ReconstructionAlgorithm):
+    """
+    Object for applying ADMM (Alternating Direction Method of Multipliers) with
+    a non-negativity constraint and a total variation (TV) prior.
+
+    Paper about ADMM: https://web.stanford.edu/~boyd/papers/pdf/admm_distr_stats.pdf
+
+    Slides about ADMM: https://web.stanford.edu/class/ee364b/lectures/admm_slides.pdf
+
+    """
+
+    def __init__(
+        self,
+        psf,
+        proj,
+        dtype=None,
+        mu1=1e-6,
+        mu2=1e-5,
+        mu3=4e-5,
+        tau=0.0001,
+        pad=False,
+        norm="backward",
+        **kwargs,
+    ):
+        """
+
+        Parameters
+        ----------
+        psf : :py:class:`~numpy.ndarray` or :py:class:`~torch.Tensor`
+            Point spread function (PSF) that models forward propagation.
+            Must be of shape (depth, height, width, channels) even if
+            depth = 1 and channels = 1. You can use :py:func:`~lensless.io.load_psf`
+            to load a PSF from a file such that it is in the correct format.
+        proj : :py:class:`function`
+            Projection function to apply at each iteration.
+        dtype : float32 or float64
+            Data type to use for optimization. Default is float32.
+        mu1 : float
+            Step size for updating primal/dual variables.
+        mu2 : float
+            Step size for updating primal/dual variables.
+        mu3 : float
+            Step size for updating primal/dual variables.
+        tau : float
+            Weight for L1 norm of `psi` applied to the image estimate.
+        pad : bool
+            Whether to pad the image with zeros before applying the PSF. Default
+            is False, as optimized data is already padded.
+        norm : str
+            Normalization to use for the convolution. Options are "forward",
+            "backward", and "ortho". Default is "backward".
+        """
+        self._mu1 = mu1
+        self._mu2 = mu2
+        self._mu3 = mu3
+        self._tau = tau
+
+        # TODO add check for proj
+        self._proj = proj
+
+        # 3D ADMM is not supported yet
+        assert len(psf.shape) == 4, "PSF must be 4D: (depth, height, width, channels)."
+        if psf.shape[0] > 1:
+            raise NotImplementedError(
+                "3D ADMM is not supported yet, use gradient descent or APGD instead."
+            )
+
+        # call reset() to initialize matrices
+        super(ADMM_PnP, self).__init__(psf, dtype, pad=pad, norm=norm, **kwargs)
+
+        # precompute_R_divmat (self._H computed by constructor with reset())
+        if self.is_torch:
+            self._R_divmat = 1.0 / (
+                self._mu1 * (torch.abs(self._convolver._Hadj * self._convolver._H))
+                + self._mu2
+                + self._mu3
+            ).type(self._complex_dtype)
+        else:
+            self._R_divmat = 1.0 / (
+                self._mu1 * (np.abs(self._convolver._Hadj * self._convolver._H))
+                + self._mu2 * np.abs(self._PsiTPsi)
+                + self._mu3
+            ).astype(self._complex_dtype)
+
+    def reset(self):
+        if self.is_torch:
+            # TODO initialize without padding
+            # initialize image estimate as [Batch, Depth, Height, Width, Channels]
+            if self._initial_est is not None:
+                self._image_est = self._initial_est
+            else:
[…]
