+++ paper/paper.md
@@ -34,8 +34,8 @@ The earliest cameras were in fact lensless, relying on long exposure times to fo
 The introduction of a lens allowed for more light throughput and therefore shorter exposure times, while retaining sharp focus. 
 The incorporation of digital sensors readily enabled the use of computational imaging techniques to post-process and enhance raw images (e.g. via deblurring, inpainting, denoising, sharpening).
 Recently, imaging scientists have started leveraging computational imaging as an integral part of lensless imaging systems, allowing them to form viewable images from the highly multiplexed raw measurements of lensless cameras, see [@boominathan2022recent] and references therein for a comprehensive treatment of lensless imaging.
-This represents a real paradigm shift in camera system design as there is more flexibility to cater the hardware to the application at hand (e.g. lightweight or flat designs).
-This increased flexibility comes however at the price of a more demanding post-processing of the raw digital recordings and a tighter integration of sensing and computation, often difficult to achieve in practice due to inefficient interactions between the various communities of scientists involved. 
+This represents a real paradigm shift in camera system design as there is more flexibility to cater the hardware to the application at hand (e.g., lightweight or flat designs).
+This increased flexibility comes, however, at the price of a more demanding post-processing of the raw digital recordings and a tighter integration of sensing and computation, often difficult to achieve in practice due to inefficient interactions between the various communities of scientists involved. 
 With `LenslessPiCam`, we provide an easily accessible hardware and software framework to enable researchers, hobbyists, and students to implement and explore practical and computational aspects of lensless imaging.
 
 # Statement of need
@@ -43,11 +43,11 @@ With `LenslessPiCam`, we provide an easily accessible hardware and software fram
 Being at the interface of hardware, software, and algorithm design, the field of lensless imaging necessitates a broad array of competences that might deter newcomers to the field.
 The purpose of `LenslessPiCam` is to provide a complete toolkit with low-cost, accessible hardware designs and open-source software, to quickly enable the exploration of novel ideas for hardware, software, and algorithm design.
 
-The DiffuserCam tutorial [@diffusercam] served as a great starting point to the present toolkit as it demonstrates that a working lensless camera can be built with cheap hardware: a Raspberry Pi, the Camera Module 2,[^1] and double-sided tape.
+The DiffuserCam tutorial [@diffusercam] served as a great starting point to the present toolkit as it demonstrates that a working lensless camera can be built with cheap hardware: a Raspberry Pi, the [Camera Module 2](https://www.raspberrypi.com/products/camera-module-v2), and double-sided tape.
 The authors also provide Python implementations of two image reconstruction  algorithms: variants of gradient descent (GD) with a non-negativity constraint; and the alternating direction method of multipliers (ADMM) [@boyd2011distributed] with an additional total variation (TV) prior.
 
 The resolution of the reconstructed images for the DiffuserCam tutorial is poor and the processing pipeline is limited to grayscale reconstruction. 
-With `LenslessPiCam`, we improve the reconstruction by using the newer HQ camera [^2] as well as a more versatile and generic RGB computational imaging pipeline. 
[…]
