+++ lensless/mask.py
@@ -111,6 +111,7 @@ class CodedAperture(Mask):
 
         super().__init__(**kwargs)
 
+
     def create_mask(self):
         """
         Creating coded aperture mask using either the MURA of MLS method
@@ -132,8 +133,8 @@ class CodedAperture(Mask):
             upscale_factor_height = self.sensor_size_px[0] / self.mask.shape[0]
             upscale_factor_width = self.sensor_size_px[1] / self.mask.shape[1]
             upscaled_mask = zoom(self.mask, (upscale_factor_height, upscale_factor_width))
-            self.mask = np.clip(upscaled_mask, 0, 1)
-        
+            upscaled_mask = np.clip(upscaled_mask, 0, 1)
+            self.mask = np.round(upscaled_mask)
 
 
     def is_prime(self, n):
