+++ paper/paper.md
@@ -33,7 +33,7 @@ Lensless imaging seeks to replace or remove the lens in a conventional imaging s
 The earliest cameras were in fact lensless, relying on long exposure times to form images on the other end of a small aperture in a darkened room/container (*camera obscura*). 
 The introduction of a lens allowed for more light throughput and therefore shorter exposure times, while retaining sharp focus. 
 The incorporation of digital sensors readily enabled the use of computational imaging techniques to post-process and enhance raw images (e.g. via deblurring, inpainting, denoising, sharpening).
-Recently, imaging scientists have started leveraging computational imaging as an integral part of lensless imaging systems, allowing them to form viewable images from the highly multiplexed raw measurements of lensless cameras, see [@boominathan2022recent] and references therein for a comprehensive treatment of lensless imaging.
+Recently, imaging scientists have started leveraging computational imaging as an integral part of lensless imaging systems, allowing them to form viewable images from the highly multiplexed raw measurements of lensless cameras (see [@boominathan2022recent] and references therein for a comprehensive treatment of lensless imaging).
 This represents a real paradigm shift in camera system design as there is more flexibility to cater the hardware to the application at hand (e.g., lightweight or flat designs).
 This increased flexibility comes, however, at the price of a more demanding post-processing of the raw digital recordings and a tighter integration of sensing and computation, often difficult to achieve in practice due to inefficient interactions between the various communities of scientists involved. 
 With `LenslessPiCam`, we provide an easily accessible hardware and software framework to enable researchers, hobbyists, and students to implement and explore practical and computational aspects of lensless imaging.
@@ -61,7 +61,7 @@ PyTorch [@pytorch] support also enables the use of GPUs for faster reconstructio
 `LenslessPiCam` is designed to be used by researchers, hobbyists, and students. In the past, we have found such open-source hardware and software platforms to be a valuable resource for researchers [@bezzam2017hardware] and students alike [@bezzam2019teaching]. Moreover, we have used `LenslessPiCam` in our own work for performing measurements, simulating data, and image reconstruction [@bezzam2022privacy], and in our graduate-level signal processing course as a [final project](https://infoscience.epfl.ch/search?ln=en&rm=&ln=en&sf=&so=d&rg=10&c=Infoscience%2FArticle&c=Infoscience%2FBook&c=Infoscience%2FChapter&c=Infoscience%2FConference&c=Infoscience%2FDataset&c=Infoscience%2FLectures&c=Infoscience%2FPatent&c=Infoscience%2FPhysical%20objects&c=Infoscience%2FPoster&c=Infoscience%2FPresentation&c=Infoscience%2FProceedings&c=Infoscience%2FReport&c=Infoscience%2FReview&c=Infoscience%2FStandard&c=Infoscience%2FStudent&c=Infoscience%2FThesis&c=Infoscience%2FWorking%20papers&c=Media&c=Other%20doctypes&c=Work%20done%20outside%20EPFL&c=&of=hb&fct__2=LCAV&p=diffusercam).
 \autoref{fig:screen_examples} demonstrates reconstructed images of our students using `LenslessPiCam` and images that were projected on a monitor 40cm away; the figure is adapted from [this report](https://infoscience.epfl.ch/record/291501?ln=en).
 
-![Reconstructions of images displayed on a monitor. "Original" is displayed on the monitor, and each subsequent row represents a different algorithm with an L2 data fidelity between the measurement and the propagated image estimate (using a measured PSF) and the following priors/regularizers on the image estimate: L2, non-negativity and total variation, non-negativity, and L1.](screen_examples.png){#fig:screen_examples width=100%}
[…]
