### 2023-06-07 9f5e73e
Paper edits for JOSS review (#53)
* Incorporate Dana's comments.

* Add extra dimension for grayscale conversion.

* Add better comparison picture with DiffuserCam.

* Update text.

* Update text.

* Small edits.
