### 2023-06-07 #4747 @ebezzam
@danasolav, thank you for the detailed and very helpful comments! I've addressed everything in the above generated PDF. Please let me know if I missed anything.

Below are some comments on a few of your points:

> Line 40: please add more information on the comparison shown in Figure 1. They don't seem to compare a reconstruction of the same image, so what is the significance of this comparison?

The purpose is to compare:
- the DiffuserCam tutorial (Biscarrat et al., 2018) which we reimplemented, and
- the camera of our toolkit

to show that reconstructions aren’t as good and limited to grayscale. I’ve done a new measurement with DiffuserCam so that the image is the same for both cameras in Figure 1.

> Line 95: the function name does not compile properly and extends beyond the line. Could you force a manual new line?

In the Docker compiled version (without line numbers) it renders correctly, [example](https://drive.google.com/file/d/1KcFaD5VPGybKm1DQAoLPdP-HVqYcNjh6/view?usp=sharing). Could it be an artifact from the peer-reviewed version with line numbers?

> Line 143: it would be helpful to explain the meaning of these numbers and their limits, where applicable.

I’ve added more description about each metric, their limits, and links/references. In 154-160, I’ve added an interpretation of Figure 6 and Table 2, which motivates the next section on using measured / simulated data.
