### 2023-06-07 204270b
Fix aspect ratio.
