+++ scripts/demo.py
@@ -1,0 +1,237 @@
+import os
+import subprocess
+import numpy as np
+import time
+from pprint import pprint
+from lensless.plot import plot_image, pixel_histogram
+from lensless.io import load_image, load_psf
+from lensless.util import resize
+import cv2
+import matplotlib.pyplot as plt
+from lensless import FISTA
+
+
+username = "pi2"
+hostname = "128.179.193.201"
+gamma = 2.2
+
+original_fp = "data/original/thumbs_up.png"
+exp = 0.02
+brightness = 90
+
+original_fp = "data/original/tree.png"
+exp = 0.24
+brightness = 100
+
+# -- display parameters
+REMOTE_PYTHON = "~/LenslessPiCam/lensless_env/bin/python"
+REMOTE_IMAGE_PREP_SCRIPT = "~/LenslessPiCam/scripts/prep_display_image.py"
+REMOTE_DISPLAY_PATH = "~/LenslessPiCam_display/test.png"
+REMOTE_TMP_PATH = "~/tmp_display.png"
+screen_res = np.array((1920, 1200))  # width, height
+pad = 0
+vshift = 0
+hshift = 0
+
+
+# -- capture parameters
+pic_delay = 2
+SENSOR_MODES = [
+    "off",
+    "auto",
+    "sunlight",
+    "cloudy",
+    "shade",
+    "tungsten",
+    "fluorescent",
+    "incandescent",
+    "flash",
+    "horizon",
+]
+REMOTE_CAPTURE_FP = "~/LenslessPiCam/scripts/on_device_capture.py"
+iso = 100
+config_pause = 2
+sensor_mode = "0"
+nbits_out = 12
+nbits = 12
+legacy = True
+bayer = True
+rgb = False
+gray = False
+down = None
+raw_data_fn = "raw_data"
+
+
+# -- reconstruction parameters
+red_gain = 1.9
+blue_gain = 1.2
+psf_fp = "data/psf/tape_rgb_31032023.png"
+downsample = 4
+dtype = "float32"
+use_torch = True
+torch_device = "cuda:0"
+disp_iter = 50
+
+# ---- fista
+n_iter = 300
+tk = 1
+
+
+# 1) Copy file to Raspbery Pi
+print("\nCopying over picture...")
+os.system('scp %s "%s@%s:%s" ' % (original_fp, username, hostname, REMOTE_TMP_PATH))
+
+prep_command = f"{REMOTE_PYTHON} {REMOTE_IMAGE_PREP_SCRIPT} --fp {REMOTE_TMP_PATH} \
+    --pad {pad} --vshift {vshift} --hshift {hshift} --screen_res {screen_res[0]} {screen_res[1]} \
+    --brightness {brightness} --output_path {REMOTE_DISPLAY_PATH} "
+print(f"COMMAND : {prep_command}")
+subprocess.Popen(
+    ["ssh", "%s@%s" % (username, hostname), prep_command],
+    shell=False,
+)
+
+# 2) Take picture
+time.sleep(pic_delay)  # for picture to display
+print("\nTaking picture...")
+
+remote_fn = "remote_capture"
+pic_command = (
+    f"{REMOTE_PYTHON} {REMOTE_CAPTURE_FP} --fn {remote_fn} --exp {exp} --iso {iso} "
+    f"--config_pause {config_pause} --sensor_mode {sensor_mode} --nbits_out {nbits_out}"
+)
+if nbits > 8:
+    pic_command += " --sixteen"
+if rgb:
+    pic_command += " --rgb"
+if legacy:
+    pic_command += " --legacy"
+if gray:
+    pic_command += " --gray"
+if down:
+    pic_command += f" --down {down}"
[…]
