### 2023-06-08 2c045bb
squared amplitude PSF
