### 2023-06-08 b7a588d
making FZA mask binary
