### 2023-06-13 #4747 @ebezzam
Hi @danasolav, just wondering if you had time to look at the changes I made and if they satisfy your points? We'll be presenting LenslessPiCam as demo at a conference next week, and would be great (if possible) to have it published by then. Thanks!
