### 2023-06-14 #4747 @ebezzam
@danasolav, @raolivei13, @siddiquesalman thank you for your input and taking the time to review the work!

@danasolav here is the DOI: [10.5281/zenodo.8036869](https://doi.org/10.5281/zenodo.8036869)
The version is 1.0.4, which can be found on [GitHub](https://github.com/LCAV/LenslessPiCam/releases/tag/v1.0.4) and installed from [PyPI](https://pypi.org/project/lensless/1.0.4).
