### 2023-06-26 #4747 @Kevin-Mattheus-Moerman
@ebezzam I've checked the repository, the paper, the archive link, and this review issue. All seems in order and I will now proceed to process this work for acceptance in JOSS.
